### 2020-01-31 #367 @uyedaj
Submission: treedata.table
Submitting Author: Josef C Uyeda (@uyedaj)
Repository:  https://github.com/uyedaj/treedata.table
Version submitted:   0.1.0
Version accepted: TBD

---

```
Package: treedata.table
Title: A Wrapper For data.table For Fast Manipulation Of Phylogenetic Trees Matched To Data
Version: 0.1.0
    person("Josef Uyeda", email= "juyeda@vt.edu", role=c("aut", "cre")),
    person("Cristian Roman-Palacios", email= "cromanpa94@email.arizona.edu", role=c("aut")),
    person("April Wright", email= "april.wright@southeastern.edu", role=c("aut"))
  )
Maintainer: Cristian Roman-Palacios <cromanpa94@email.arizona.edu>
Description: This package combines trait data and a phylogenetic tree (or trees) into a
    single object of class treedata.table. The resulting object can be easily
    manipulated to simultaneously change the trait- and tree-level sampling.
    Currently implemented functions allow users to use a data.table syntax when
    performing operations on the trait dataset within the treedata.table object.
  data.table, R (>= 2.10)
  lazyeval, ape, geiger, utils
```

The package allows manipulation of data frames with `data.table` while maintaining a matched ordering to an `ape` phylogenetic tree. This allows rapid data munging without worrying about a
mismatch with the tree order.

Researchers doing phylogenetic analyses.

`treeplyr` works similarly but is built on the package `dplyr` rather than `data.table`, which has some advantages to `dplyr`.
