### 2020-02-19 #367 @jooolia
Hi @uyedaj!
Thanks for the submission and for your work on this package. I will be your editor for this submission. I have done a few of the initial editor checks.

---

- For the licence, the description says MIT, but the LICENCE file is not complete. This needs to be completed.
- I found the repository to be https://github.com/uyedaj/treedata.table, could you add the link at the top of the submission info so that it is easier to find the repo?
- Spell-checking: Looks good.
- Below are some automated checks performed on the package. Please address those without the tag *reviewers*.

[goodpractice::gp](https://github.com/MangoTheCat/goodpractice/) output
     ── GP treedata.table

    ✖ write unit tests for all functions, and all package code in
     general. 34% of code lines are covered by test cases.

     R/as.treedata.table.R:40:NA
     R/as.treedata.table.R:43:NA
     R/as.treedata.table.R:44:NA
     R/as.treedata.table.R:49:NA
     R/as.treedata.table.R:55:NA

- [ ] *reviewers* will determine if this appropriate amount of test coverage

- [ ] not use "Depends" in DESCRIPTION, as it can cause name clashes, and
     poor interaction with other packages. Use "Imports" instead.
This can be argued if necessary. For more info see http://r-pkgs.had.co.nz/description.html

- [ ] add a "URL" field to DESCRIPTION. It helps users find information
     about your package online. If your package does not have a
     homepage, add an URL to GitHub, or the CRAN package package page.

- [ ] add a "BugReports" field to DESCRIPTION, and point it to a bug
     tracker. Many online code hosting services provide bug trackers for
     free, https://github.com, https://gitlab.com, etc.

- [ ] use '<-' for assignment instead of '='. '<-' is the standard, and R
     users and developers are used it and it is easier to read your code
     for them if you use '<-'.

      R/detectCharacterType.R:76:8
      R/detectCharacterType.R:104:8
      R/extractVector.R:31:10
      R/tdt_methods.R:16:5

- [ ] avoid long code lines, it is bad for readability. Also, many people
     prefer editor windows that are about 80 characters wide. Try make
     your lines shorter than 80 characters

      R/as.treedata.table.R:1:1
      R/as.treedata.table.R:3:1
      R/as.treedata.table.R:4:1
      R/as.treedata.table.R:5:1
      R/as.treedata.table.R:11:1

- [ ] avoid sapply(), it is not type safe. It might return a vector, or a
     list, depending on the input data. Consider using vapply() instead.

      R/as.treedata.table.R:71:16
      R/as.treedata.table.R:118:8
      R/pull.treedata.table.R:20:18

- [ ] avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and
     1:NCOL(...) expressions. They are error prone and result 1:0 if the
     expression on the right hand side is zero. Use seq_len() or
     seq_along() instead.

      R/as.treedata.table.R:59:38
      R/treedata.table.R:67:33
      R/treedata.table.R:69:47

- [ ] *reviewers*  to consider and give feedback on
      ✖ fix this R CMD check WARNING: LaTeX errors when creating PDF
      version. This typically indicates Rd problems. LaTeX errors found:
      !pdfTeX error: pdflatex (file ts1-zi4r): Font ts1-zi4r at 600 not
      found ==> Fatal error occurred, no output PDF file produced!

- [ ] avoid 'T' and 'F', as they are just variables which are set to the
     logicals 'TRUE' and 'FALSE' by default, but are not reserved words
     and hence can be overwritten by the user.  Hence, one should always
     use 'TRUE' and 'FALSE' for the logicals.

      R/as.treedata.table.R:NA:NA
      R/as.treedata.table.R:NA:NA
      R/tdt_methods.R:NA:NA

 ─────────────────────────────────────────────────────────────

---

I will look for reviewers after these small issues have been addressed. Feel free to let me know if you have any questions.
 Thanks! Julia
