### 2020-03-03 7764ef2
addressing R Open Sci initial checks
