+++ R/pull.treedata.table.R
@@ -17,7 +17,7 @@ pull.treedata.table <- function(tdObject, type = "dat") {
   }
   full <-
     if (type == "dat") {
-      matches <- sapply(tdObject$dat, function(x) sum(x %in% tdObject$phy$tip.label))
+      matches <- vapply(tdObject$dat, function(x) sum(x %in% tdObject$phy$tip.label), integer(1))
       if(any(matches==nrow(tdObject$dat))){
         tdObject$dat
       } else{
