### 2020-03-10 #367 @wrightaprilm
I think between my commit (7764ef2) and Scott's edit of the initial PR, we should have all these notes handled!
