### 2020-06-20 #367 @jooolia
Hi @wrightaprilm  and @uyedaj,
We are now looking for reviewers. Please add the rOpenSci review badge to your README.
You can do this with by adding to your README:

Thanks! Julia
