### 2020-07-23 #367 @Bisaloo
- [ ] **Performance:** Any performance claims of the software been confirmed. -> not tested

Estimated hours spent reviewing: 2h30 + 1h30 + 1h

---

I've added some (*optional*) comments. These are not necessary for your package to follow the **current** best practices but standards are always evolving and I believe addressing these now will save you some trouble later down the road.

It would be super useful for me if you could use a different commit for each the points raised in my review :+1:.

I hope everything makes sense. Please let me know if something is not clear.

#### DESCRIPTION and other packaging files

I recommend you have a look at the [packaging guidelines in rOpenSci's development guide](https://devguide.ropensci.org/building.html). It lists a number of points that are currently missing from your package. Among others:

  - [ ] It's recommended practice to add an ORCiD for the authors if they have one. [Source](https://ropensci.org/technotes/2018/10/08/orcid/)
  - [ ] Please add a contributing guide.
  - [ ] There is a mismatch in the maintainer according to the `Authors@R` and the `Maintainer` fields in the `DESCRIPTION`. It's recommended to leave out the `Maintainer` field entirely and it will be automatically generated from `Authors@R`.

- As far as I know, the `Type` field in the `DESCRIPTION` is not standard and in all cases unnecessary

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/DESCRIPTION#L2

- The package `Title` is not actually using title case. I don't care much personally but in my experience, this can cause issues when you submit to CRAN

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/DESCRIPTION#L3

You can verify this with the `tools::toTitleCase()` function.

- [ ] Please avoid as much as possible to use `Depends` when you can use `Imports`. Here is a relevant quote from Hadley Wickham and Jenny Bryan's book on R package development:

> Depends: Prior to the rollout of namespaces in R 2.14.0, Depends was the only way to “depend” on another package. Now, despite the name, you should almost always use Imports, not Depends. You’ll learn why, and when you should still use Depends, in namespaces.

- [ ] According to the [CRAN guidelines for MIT licenses](https://www.r-project.org/Licenses/MIT), your previous version of `LICENSE` was correct (before https://github.com/uyedaj/treedata.table/commit/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c). You just need to update to name of the copyright holder(s).

- [ ] The `README` does not contain enough information about the package, how to install and use it. You may find [this relevant chapter](https://devguide.ropensci.org/building.html#readme) of rOpenSci's devguide useful.
- [ ] Since this package is presented as addressing a performance need, it would be very nice to see some benchmarks to back this up. You can use the [microbenchmark package](https://cran.r-project.org/package=microbenchmark) for this.
- (*optional*) it can be nice to add more badges to your `README` to advertise the fact that you follow the current best practices in package development. In particular, I like badges with code coverage. [Here are the instructions on how to do this with codecov](https://stackoverflow.com/questions/54010651/codecov-io-badge-in-github-readme-md).

- [ ] Please avoid multiline instructions (separated by `;`). It makes the code harder the read. This will also help with the long lines warning reported by goodpratice in @jooolia's comment:

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/as.treedata.table.R#L118

- [ ] When you perform tests on objects that are already booleans (`TRUE`/`FALSE`), you don't need to write `== TRUE`. E.g.:

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/as.treedata.table.R#L44

```r
if (!equal_T) { ... }
```

- [ ] Class testing should be done with `inherits()` (for S3 objects, which is the case for ape objects), or `is()` (for S4 objects). Please see [this blog post by Martin Maechler](https://developer.r-project.org/Blog/public/2019/11/09/when-you-think-class.-think-again/index.html) for more info. Here is a concrete example of an issue caused by this in your package:

```r
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
p <- pull.treedata.table(td, type = "phy")
d <- pull.treedata.table(td, type = "dat")
td2 <- as.treedata.table(p, d)
# You would expect to get identical(td, td2) but this line actually throws an error
```

- [ ] Regarding your replacement of `1:length()` by `seq_len()` in https://github.com/uyedaj/treedata.table/commit/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c, the correct syntax of `seq_len()` doesn't use `1:`. E.g.,

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/as.treedata.table.R#L59

should read

```r
colnames(data) <- paste("trait", seq_len(length(data)), sep="")
```

or better

```r
colnames(data) <- paste("trait", seq_along(data), sep="")
```

- [ ] Communicate information to users via the `message()` function instead of `cat()`

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/as.treedata.table.R#L83

This will always surely cause issues upon CRAN submission if not corrected. See this [detailed SO answer](https://stackoverflow.com/a/36700294) for more info about the differences between `message()` and `cat()`.

- [ ] On a related note, you don't need to use `paste0()` in `message()` calls. The `message()` will automatically concatenate the arguments you give it:

```r
message("this ", "is ", "a ", "test!")
#> this is a test!
```

- [ ] There is a problem in `as.treedata.table()` with the `name_column` argument not working. No matter what the user enters in `name_column` or what the auto-detection code finds, you always use the first column for tip.labels:

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/as.treedata.table.R#L84-L90

Reproducible example showing the bug:

``` r
library(treedata.table)
#> Loading required package: lazyeval
#> Loading required package: ape
#> Loading required package: geiger
#> Loading required package: data.table

data(anolis)

tre <- anolis$phy
dat <- anolis$dat

# shuffle columns
dat <- dat[, sample(ncol(dat), ncol(dat))]

as.treedata.table(tre, dat)
#> Phylo object detected
#> 100 tip(s) dropped from the original tree
#> 100 tip(s)  dropped from the original dataset
#> Warning in ape::drop.tip(tree, tree_not_data): drop all tips of the tree:
#> returning NULL
#> $phy
#> 0 phylogenetic tree
#>
#> $dat
#> Empty data.table (0 rows and 11 cols): PCIV_lamella_num,tip.label,island,hostility,ecomorph,PCI_limbs...
```

This behaviour should be tested in tests.

- (*optional*) it would be nice to have a message indicating which column was auto-detected as containing the tip.labels when `name_column = "detect"` in `as.treedata.table()`

- (*optional*) uniformise the indents in your code. You sometimes use tabs and sometimes spaces. This can be fixed by running `styler::style_pkg()` in the root of your RStudio project.

- [ ] As far as I can tell, the `repeatsAsDiscrete` is not used in `detectCharacterType()` or `detectAllCharacters()`

- [ ] This line will cause wrong identifications since the conversion of a data.frame to a matrix forces all elements to the same type.

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/detectCharacterType.R#L34

This can be seen by using the two examples you provide

```r
data(anolis)

detectCharacterType(anolis$dat[,1])
#> [1] "discrete"

detectAllCharacters(anolis$dat)
#> [1] "continuous" "continuous" "continuous" "continuous" "continuous" "continuous" "continuous" "continuous"
#> [9] "continuous" "discrete"   "discrete"

detectCharacterType(anolis$dat[,1]) == detectAllCharacters(anolis$dat)[1]
#> FALSE
```

Regarding my later point about tests, this would make a good test: ensuring that you find the correct number of discrete/continuous characters in the anolis dataset.

- [ ] I think you missed some `1:` instances from @jooolia's comment. E.g.,

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/detectCharacterType.R#L108

- [ ] It may be my unfamiliarity with data.table but I don't understand the warning/prompt in `droptreedata.table()`. As far as I can tell, the original data is NOT modified. Using your example:

```r
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
td_old <- td
td_new <- droptreedata.table(tdObject = td, taxa = c("chamaeleonides" ,"eugenegrahami"))

identical(td, td_old)
#> TRUE

identical(td, td_new)
#> FALSE
```

- [ ] The dots handling in `extractVector()` may probably be simplified with lazyeval since you already have it as a dependency

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/extractVector.R#L22

- [ ] It would be useful to add a `match.arg(type, c("dat", "phy"))` in `pull.treedata.table()`

- [ ] In `tdt`, do you really need `...`? Wouldn't an `FUN` arg be sufficient?

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/tdt.R#L4

- [ ] Since you provide a `head()` method for `treedata.table`, it would be nice to have a `tail()` method as well

- [ ] `paste()`s are unnecessary here since `cat()` will already do this:

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/tdt_methods.R#L62-L74

- [ ] The output of `summary.treedata.table()` is slightly confusing in my opinion. On the example you provide (with anolis), where no taxa are dropped from the tree or the data, you get:

> These taxa were dropped from the tree: OK
> These taxa were dropped from the data: OK

It's not obvious what's this OK is supposed to mean, I think it would be simpler if it was empty in this case.

- [ ] When I run examples from `[.treedata.table()`, I get the following warning:

> Warning message:
In 1:seq_along(x$dat) :
  numerical expression has 11 elements: only the first used

Could you look into it please? I assume you meant `seq_along(x$dat)`

- (*optional*) your tests are not compatible with the [upcoming v3 of the testthat package](https://github.com/r-lib/testthat/blob/master/NEWS.md#testthat-development-version). In particular, you could replace the (soon to be) deprecated `expect_is()` function by `expect_s3_class()`. Excepted these two lines

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/tests/testthat/test.treedata.table.R#L62-L63

that need to use `expect_type()`

- [ ] I think the test coverage needs to be increased a lot, as shown by the bugs uncovered during this review. I sometimes noted what would be good candidates for tests. As a good starting point, you can also look at which lines are not covered by your tests on codecov: https://codecov.io/github/uyedaj/treedata.table?branch=master. I see no technical limitations that would prevent you to reach 100% coverage for this package but for now, I think you should at least try to reach 80% coverage.

- (*optional*) you may want to use markdown syntax in your Roxygen comments. This produces more readable documentation in the source file in my opinion. Automatic conversion of your current Roxygen comments to markdown can be done with the [roxygen2md](https://roxygen2md.r-lib.org/) package.
- Please add a sentence for the `name_column` argument to explain that `"detect"` (the default) will auto-detect this column:

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/as.treedata.table.R#L13

- In `as.treedata.table()`, you declare

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/as.treedata.table.R#L10

but this is unnecessary since you don't use `setDT()` here and `as.data.table()` is prefixed with the `data.table::` namespace

- I'm not sure what the first word means in your `@return` roxygen comments. I think it will be less confusing if you remove these(?)

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/as.treedata.table.R#L14

- (*optional*) you can use the `@inheritParams detectCharacterType` roxygen comment in the documentation of `detectAllCharacters()` to avoid duplicating the documentation of the function arguments. This is useful because it ensures the documentation of these functions will always stay in sync in the future. No risk of forgetting to update one of the two!

- (*optional*) I find it useful to explicitly state which is default for all arguments (when it exists) in the documentation. E.g., in

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/detectCharacterType.R#L47

I would change it to something like

```r
#' @param returnType Either discrete (the default) or continuous
```

- [ ] Please expand the documentation of the `hasNames()`/`forceNames()` functions by providing a short explanation of the function purpose and/or a `@return` roxygen comment.

- [ ] The documentation fo `extractVector()` should be updated to indicate that multiple column names can be passed to `...`, which means you don't necessarily get a named vector but a list of named vectors, as opposed to what the documentation says:

```r
extractVector(td, "SVL", "island")
```

- [ ] I think the function name `pull.treedata.table` is slightly confusing as it sounds like a S3 and it's not. Something along the lines of `pull_treedata.table()` or `pulldata.data.table()` (you already have a `dropdatatree.table()` function so this would make sense) or ??? would probably be better.

- [ ] Shouldn't this just be "If negative, all but the n last rows of x" (i.e., remove "first"):

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/tdt_methods.R#L4-L6

- [ ] This doesn't seem correct. The dots are ignored in `head.treedata.table()` as far as I can tell

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/R/tdt_methods.R#L7

- [ ] To fix the `NOTE` in `R CMD check`, you need to `importFrom(utils,head)` before you try to define another method for this generic. This can be done by adding a `#' @importFrom utils head` roxygen comment in the roxygen chunk of `head.treedata.table()` for example.

#### Vignette

- [ ] Please update the list of authors in the vignette:

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/vignettes/treedata.table.Rmd#L3

- (*optional*) it may useful for users to mention the differences between `[[.treedata.table()` and `extractVector()`, namely that `[[.treedata.table()` has an extra `exact` argument to enable partial match while `extractVector()` can extract multiple columns and accepts non-standard evaluation.

- [ ] If you remove the "[1] FALSE FALSE" (which are actually output, and not R code) from this chunk, you will be able to remove the `eval = FALSE`. It's awlays better when all chunks run and can be copied/pasted in the console directly

https://github.com/uyedaj/treedata.table/blob/7764ef2c4a0e56f1cb5fbcef6d910fb8ba0eea8c/vignettes/treedata.table.Rmd#L102-L116

#### Misc

- I can't reproduce the other `R CMD check` error mentioned by @jooolia (related to LaTeX). It might be version specific???
