### 2020-07-30 #367 @wrightaprilm
Hey @uyedaj and @cromanpa, I'll start tackling these early next week. I'm devoting 100% of my time to class prep this week.
