### 2020-08-01 #367 @cromanpa94
Sounds perfect @wrightaprilm!! I'm also go over the @Bisaloo's comments during the weekend!
