### 2020-08-01 df722b1
Revised: Installation instructions
