### 2020-08-01 86edeb2
Revised: Community guidelines included to the README
