### 2020-08-01 0b36f8e
add libmagick++
