### 2020-08-02 79461cb
update travis libmagic
