+++ .travis.yml
@@ -13,7 +13,7 @@ notifications:
   email:
     on_success: change
     on_failure: change
-osx_image: xcode9.2
+osx_image: xcode11.4
 addons:
   apt:
     update: true
