### 2020-08-04 #367 @karinorman
Hi @cromanpa94 - I really enjoyed exploring your package. It seems like a really nice tool for making phylogeny data easier to work with and I'm sure it's going to get a lot of use! See below for my comments.

>Consider putting a link in the README.
> - The data manipulation in `as.treedata.table()` example should be explained, or you could do it behind the scenes and just import the objects with an explanation of what they are/their structure.
> - No example for `print.treedata.table`
> The maintainer denoted in `Authors@R`is defferent from `Maintainer:` in DESCRIPTION. I don't see an explaination of contribution guidlines, the ROpenSci package documentation (https://devguide.ropensci.org/collaboration.html#friendlyfiles) gives good guidance on adding it.

> Not necessarily something to be fixed, but so you're aware, I'm still running R 3.6.3 and the dependency `mnormt` will not install from CRAN unless you're running R >4.0.0, so I had to manually install from an old github version (e.g. `devtools::install_github('cran/mnormt@R-3.0.3')`) before your install would work.

   `goodpractice::gp()` output:
   - linebreaks under 80 characters
   - poor unit test coverage (only 34%)

> - Need to flesh out the README, a list of important aspects can be found in the ROpenSci package guidelines (section 1.5) https://devguide.ropensci.org/building.html.

4 hrs

---

This is my first review for ROpenSci and I tried to approach it from the perspective of a naive user, so most feedback is related to things that may be difficult with initial use. Of course you are the expert on your users and this kind of data so I'll look forward to learning more from you about what does and doesn't make sense in this context.

General Comments:
- The warnings for co-indexing calls (e.g. ` td[, head(.SD, 1), by = "ecomorph"]` from the vignette don't seem particularly informative, I would suppress unless necessary.
- I think having the user confirm the changes for `droptreedata.table` is unnecessary. For me at least it made me assume that the function was modifying an object in place, which I realized it wasn't after some exploration.

Functions:
- Consider changing the naming  convention to `*.td.table` rather than `*.treedata.table` for the sake of brevity. You could also add a `td` to other exported functions (e.g. td.extractVector).
- The `detectCharacter` functions, `filterMatrix`, `forceNames`, and `hasNames` seem like they could be helper functions that facilitate the other main functions, but maybe don't need to be exported. If it makes sense I would not export them. If not, a little more explanation of how they fit into a workflow in the vignette would be helpful.
- `forceNames` and `hasNames` could use more detailed description and/or justification, I'm not sure their functionality even after running the example. `data(anolis)` and `forceNames(anolis$dat, "row")` return seemingly the same object.
- It could be useful to make dropping tips or dataframe entries optional when matching trees/dataframes instead of automatically dropping from either to match, so that the user has the option to preserve all data.
- `filterMatrix` would be cleaner if `charType` wasn't required as an argument but instead calculated within the function. I have a hard time imagining a scenario in which you would want a vector of character types that didn't match the matrix you were already giving the function.
- I'm curious about the `pull.treedata.table` function. Is there some utitlity in having a function that mimics the `$` operator? Or maybe I'm missing some of the applications of this function?

`detectCharacter` Functions:
- Please explain the definition of continuous or discrete in the descriptions of `detectCharacterType`, `detectCharacterChanges`, and `filterMatrix`.
- I would change the language around "character" which is a specific object type, whereas this function appears to perform on multiple vector types.`detectVectorType` or `detectColumnType` may be more intuitive.
- There is strange behavior in the examples for these functions. For example `detectCharacterType(anolis$dat[,1])` returns "discrete", but `detectAllCharacters(anolis$dat[,1:3])` returns three "continuous" entries. From my understanding of how the functions work I would expect the first entry to be "discrete" to match `detectCharacterType(anolis$dat[,1])`.
