### 2020-08-04 07746cf
REVIEW:Type field deleted
