### 2020-08-04 59872e5
Fix multiline as.treedata.table
