### 2020-08-04 9afe310
REVIEW:Booleans fixed in ifelse statements
