### 2020-08-04 1b5c188
REVIEW:using inherits instead of class
