+++ R/tdt.R
@@ -31,7 +31,7 @@
 
 tdt <- function(tdObject, ...){
 
-  if(class(tdObject) != "treedata.table" ){
+  if(!inherits(tdObject, c('treedata.table')) ){
     stop("Please use a class 'treedata.table' object \n")
   }
 
@@ -43,7 +43,7 @@ tdt <- function(tdObject, ...){
   env <- new.env(parent = parent.frame(), size = 1L)
   env$dat <- tdObject$dat
 
-  if(class(tdObject$phy) == "phylo"){
+  if(inherits(tdObject$phy, c('phylo'))){
     cat("Phylo object detected. Expect a single function output")
     env$phy <- tdObject$phy
     out <- eval(call, env)
