### 2020-08-04 700e0ed
REVIEW:Fix seq_along
