### 2020-08-04 c7ebb3b
REVIEW:paste0() within message() were removed
