+++ R/tdt_methods.R
@@ -60,7 +60,7 @@ summary.treedata.table <- function(object, ...){
   }
 
   message("A treedata.table object", "\n")
-  message(paste("The dataset contains ", ncol(object$dat), " traits"),
+  message("The dataset contains ", ncol(object$dat), " traits",
       "\n")
   types <- stats::setNames(suppressWarnings(detectAllCharacters(as.matrix(object$dat))),
                     colnames(object$dat))
@@ -68,12 +68,12 @@ summary.treedata.table <- function(object, ...){
       "\n")
   message("Discrete traits: ", names(types)[which(types == "discrete")],
       "\n")
-  message(paste("The following traits have missing values:", paste(names(types)[apply(object$dat,
-                                                                                  2, function(y) any(is.na(y)))], collapse = ", "), "\n"))
-  message(paste("These taxa were dropped from the tree:", paste(attributes(object)$tree_not_data,
-                                                            collapse = ", "), "\n"))
-  message(paste("These taxa were dropped from the data:", paste(attributes(object)$data_not_tree,
-                                                            collapse = ", "), "\n"))
+  message("The following traits have missing values:", paste(names(types)[apply(object$dat,
+                                                                                  2, function(y) any(is.na(y)))], collapse = ", "), "\n")
+  message("These taxa were dropped from the tree:", paste(attributes(object)$tree_not_data,
+                                                            collapse = ", "), "\n")
+  message("These taxa were dropped from the data:", paste(attributes(object)$data_not_tree,
+                                                            collapse = ", "), "\n")
   print(object, ...)
   if( !is.null(attr(object, "modified")) ){ message("\n    This is NOT the original treedata.table object") }
 }
