### 2020-08-04 cb04cf5
REVIEW:name_column argument fixed for "detect"
