### 2020-08-04 54a8559
REVIEW:testthat updated to include correct column auto-detection
