### 2020-08-04 5b1951b
REVIEW:Message printing the auto-detected column
