+++ R/droptreedata.table.R
@@ -8,25 +8,24 @@
 #' @return An object of class \code{treedata.table} with matching \code{$dat} and \code{$phy} elements based on whether \code{taxa} were dropped or not.
 #' @examples
 #' data(anolis)
-#' #With a multiphylo object in the treedata.table object
+#' # With a multiphylo object in the treedata.table object
 #' td <- as.treedata.table(anolis$phy, anolis$dat)
-#' droptreedata.table(tdObject=td, taxa=c("chamaeleonides" ,"eugenegrahami" ))
+#' droptreedata.table(tdObject = td, taxa = c("chamaeleonides", "eugenegrahami"))
 #'
-#' #With a multiphylo object in the treedata.table object
-#' treesFM<-list(anolis$phy,anolis$phy)
+#' # With a multiphylo object in the treedata.table object
+#' treesFM <- list(anolis$phy, anolis$phy)
 #' class(treesFM) <- "multiPhylo"
 #' td <- as.treedata.table(treesFM, anolis$dat)
-#' droptreedata.table(tdObject=td, taxa=c("chamaeleonides" ,"eugenegrahami" ))
-#'
+#' droptreedata.table(tdObject = td, taxa = c("chamaeleonides", "eugenegrahami"))
 #' @export
 
 
 droptreedata.table <- function(tdObject, taxa) {
-  if(class(tdObject) != "treedata.table" ){
+  if (class(tdObject) != "treedata.table") {
     stop("Please use a class 'treedata.table' object \n")
   }
 
-  if(class(taxa) != "character" ){
+  if (class(taxa) != "character") {
     stop("Please use a class 'character' object for taxa \n")
   }
 
@@ -37,27 +36,23 @@ droptreedata.table <- function(tdObject, taxa) {
     .dat <- tdObject$dat
     .phy <- tdObject$phy
 
-    if(inherits(.phy, c('phylo')) ){
-
-      .dat <-.dat[!.phy$tip.label %in% taxa]
+    if (inherits(.phy, c("phylo"))) {
+      .dat <- .dat[!.phy$tip.label %in% taxa]
       .phy <- ape::drop.tip(.phy, which(.phy$tip.label %in% taxa))
-
-    }else{
-
+    } else {
       .dat <- .dat[!.phy[[1]]$tip.label %in% taxa]
-      .phy <- lapply(.phy,ape::drop.tip,tip=which(.phy[[1]]$tip.label %in% taxa))
-      class(.phy)<-'multiPhylo'
+      .phy <- lapply(.phy, ape::drop.tip, tip = which(.phy[[1]]$tip.label %in% taxa))
+      class(.phy) <- "multiPhylo"
     }
 
 
     tdObject$dat <- .dat
     tdObject$phy <- .phy
-    attr(tdObject,'modified') <- 1
+    attr(tdObject, "modified") <- 1
     return(tdObject)
     message("Changes were included to the ORIGINAL data")
-  } else{
+  } else {
     message("NO changes made to the ORIGINAL data")
     return(tdObject)
   }
 }
-
