### 2020-08-04 600752d
Remove new testthat
