### 2020-08-05 e00fef9
REVIEW:testthat updated to include correct column auto-detection (UPDATED)
