+++ tests/testthat/test.treedata.table.R
@@ -71,4 +71,11 @@ test_that("phy and dat objects can be extracted correctly using pull.treedata.ta
   expect_is(pull.treedata.table(td, type = "dat"), "data.table")
 })
 
-
+test_that("Column containing tip labs can be correctly detected", {
+  tre <- anolis$phy
+  dat1 <- anolis$dat
+  dat2 <- dat1[, sample(ncol(dat1), ncol(dat1))]
+  td1 <- as.treedata.table(tre, dat1)
+  td2 <- as.treedata.table(tre, dat2)
+  expect_equal(td1$phy, td2$phy)
+})
