### 2020-08-05 41f32f5
Review:remove repeatsAsDiscrete from detectCharacterType() and detectAllCharacters()
