### 2020-08-05 099a0a8
REVIEW:removed as.matrix() from detectAllCharacters
