### 2020-08-05 4752e44
REVIEW:Test if the correct number of discrete/continuous characters are being found in the anolis dataset
