+++ R/detectCharacterType.R
@@ -105,14 +105,13 @@ forceNames <- function(dat, nameType = "row") {
   if (nType == "row" | nType == "rowcol") {
     if (!hasNames(dat, nameType = "row")) {
       nrows <- dim(dat)[1]
-      rownames(dat) <- paste("n", 1:nrows, sep = "")
+      rownames(dat) <- paste("n", seq_along(dat[,1]), sep = "")
     }
   }
   if (nType == "col" | nType == "rowcol") {
     if (!hasNames(dat, nameType = "col")) {
       ncols <- dim(dat)[2]
-      colnames(dat) <- paste("n", 1:ncols, sep = "")
-    }
+      rownames(dat) <- paste("n", seq_along(dat[1,]), sep = "")     }
   }
 
   dat
