### 2020-08-05 8d4b9e1
REVIEW:Remove as.matrix from detectAllCharacters()
