### 2020-08-05 19a777c
Using seq_len instead of 1:
