### 2020-08-05 3cd17b9
REVIEW:Test if the correct number of discrete/continuous characters
