### 2020-08-05 a6d0b60
REVIEW:promt removed in droptreedata.table
