### 2020-08-05 23cfafe
remove extra line in summary.treedata.table()
