### 2020-08-05 63a2935
REVIEW:extractVector simplified for ...
