+++ man/tail.treedata.table.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/tdt_methods.R
-\name{head.treedata.table}
-\alias{head.treedata.table}
-\title{Return the first part of an treedata.table object}
+\name{tail.treedata.table}
+\alias{tail.treedata.table}
+\title{Return the last part of an treedata.table object}
 \usage{
-\method{head}{treedata.table}(x, n = 6L, ...)
+\method{tail}{treedata.table}(x, n = 6L, ...)
 }
 \arguments{
 \item{x}{a treedata.table object}
@@ -14,10 +14,10 @@
 \item{...}{Additional arguments passed to head.data.table}
 }
 \description{
-Return the first part of an treedata.table object
+Return the last part of an treedata.table object
 }
 \examples{
 data(anolis)
 td <- as.treedata.table(anolis$phy, anolis$dat)
-head(td)
+tail(td)
 }
