+++ man/forceNames.Rd
@@ -14,7 +14,7 @@ forceNames(dat, nameType = "row")
     \item{"row"}{Rows}
          \item{"col"}{Columns}
          \item{"rowcol"}{Both rows and columns}
-}}
+    }}
 }
 \description{
 Force names for rows, columns, or both
