+++ man/tdt.Rd
@@ -26,18 +26,17 @@ This function allows R functions that use trees and data to be run on
 data(anolis)
 \dontrun{
 
-#A treedata.table object with a phylo $phy
+# A treedata.table object with a phylo $phy
 td <- as.treedata.table(anolis$phy, anolis$dat)
-tdt(td, geiger::fitContinuous(phy, extractVector(td, SVL), model="BM", ncores=1))
-tdt(td, phytools::phenogram(phy, extractVector(td, SVL), quiet=TRUE, spread.labels=FALSE))
+tdt(td, geiger::fitContinuous(phy, extractVector(td, SVL), model = "BM", ncores = 1))
+tdt(td, phytools::phenogram(phy, extractVector(td, SVL), quiet = TRUE, spread.labels = FALSE))
 
 
-#A treedata.table object with a multiPhylo $phy
-treesFM<-list(anolis$phy,anolis$phy)
+# A treedata.table object with a multiPhylo $phy
+treesFM <- list(anolis$phy, anolis$phy)
 class(treesFM) <- "multiPhylo"
 td <- as.treedata.table(treesFM, anolis$dat)
-tdt(td, geiger::fitContinuous(phy, extractVector(td, SVL), model="BM", ncores=1))
-tdt(td, phytools::phenogram(phy, extractVector(td, SVL), quiet=TRUE, spread.labels=FALSE))
-
+tdt(td, geiger::fitContinuous(phy, extractVector(td, SVL), model = "BM", ncores = 1))
+tdt(td, phytools::phenogram(phy, extractVector(td, SVL), quiet = TRUE, spread.labels = FALSE))
 }
 }
