### 2020-08-05 42491d2
extractVector examples updated
