### 2020-08-05 0f31ded
Update extractVector
