+++ R/extractVector.R
@@ -9,8 +9,10 @@
 #'
 #' data(anolis)
 #' td <- as.treedata.table(tree = anolis$phy, data = anolis$dat)
-#' extractVector(td, "SVL") # extracts the named vector for SVL from the td object
-#' extractVector(td, "SVL","ecomorph") # extracts the named vector for SVL and ecomorph from the td object
+#' # extracts the named vector for SVL from the td object
+#' extractVector(td, "SVL")
+#' # extracts the named vector for SVL and ecomorph from the td object
+#' extractVector(td, "SVL","ecomorph")
 #' @export
 
 extractVector <- function(tdObject, ...) {
