+++ man/extractVector.Rd
@@ -21,5 +21,8 @@ Returning a named vector from a treedata.table object
 
 data(anolis)
 td <- as.treedata.table(tree = anolis$phy, data = anolis$dat)
-extractVector(td, "SVL") # extracts the named vector for SVL from the td object
+# extracts the named vector for SVL from the td object
+extractVector(td, "SVL")
+# extracts the named vector for SVL and ecomorph from the td object
+extractVector(td, "SVL","ecomorph")
 }
