### 2020-08-05 292295c
S3 methods updated
