+++ NAMESPACE
@@ -2,10 +2,8 @@
 
 S3method("[",treedata.table)
 S3method("[[",treedata.table)
-S3method(head,treedata.table)
 S3method(print,treedata.table)
 S3method(summary,treedata.table)
-S3method(tail,treedata.table)
 export(as.treedata.table)
 export(detectAllCharacters)
 export(detectCharacterType)
@@ -14,7 +12,9 @@ export(extractVector)
 export(filterMatrix)
 export(forceNames)
 export(hasNames)
+export(head.treedata.table)
 export(pull.treedata.table)
+export(tail.treedata.table)
 export(tdt)
 importFrom(data.table,as.data.table)
 importFrom(data.table,setDT)
