+++ NAMESPACE
@@ -2,8 +2,10 @@
 
 S3method("[",treedata.table)
 S3method("[[",treedata.table)
+S3method(head,treedata.table)
 S3method(print,treedata.table)
 S3method(summary,treedata.table)
+S3method(tail,treedata.table)
 export(as.treedata.table)
 export(detectAllCharacters)
 export(detectCharacterType)
@@ -12,9 +14,12 @@ export(extractVector)
 export(filterMatrix)
 export(forceNames)
 export(hasNames)
-export(head.treedata.table)
 export(pull.treedata.table)
-export(tail.treedata.table)
 export(tdt)
+importFrom(ape,drop.tip)
 importFrom(data.table,as.data.table)
 importFrom(data.table,setDT)
+importFrom(geiger,name.check)
+importFrom(lazyeval,lazy_dots)
+importFrom(utils,head)
+importFrom(utils,tail)
