+++ man/anolis.Rd
@@ -4,7 +4,9 @@
 \name{anolis}
 \alias{anolis}
 \title{Anole data}
-\format{An object of class \code{list} of length 2.}
+\format{
+An object of class \code{list} of length 2.
+}
 \usage{
 data(anolis)
 }
