### 2020-08-06 17e940c
Additional tests
