+++ R/pull.treedata.table.R
@@ -3,7 +3,7 @@
 #' @param tdObject A treedata.table object
 #' @param type Whether the output of the function is a tree ('type="phylo"') or a data.table ('type="dat"')
 #'
-#' @return A \code{data.table} or \code{phylo} object from the original \code{treedata.table} object
+#' @return A `data.table` or `phylo` object from the original `treedata.table` object
 #' @examples
 #' data(anolis)
 #' td <- as.treedata.table(anolis$phy, anolis$dat)
