+++ R/droptreedata.table.R
@@ -1,11 +1,11 @@
-#' Function dropping taxa from an object of class \code{treedata.table}
+#' Function dropping taxa from an object of class `treedata.table`
 #'
-#' This function can be used to remove species from an object of class \code{treedata.table}. The resulting \code{treedata.table} will
-#' include fully matching \code{$dat} and \code{$phy} elements. The user should confirm the changes before they are processed.
+#' This function can be used to remove species from an object of class `treedata.table`. The resulting `treedata.table` will
+#' include fully matching `$dat` and `$phy` elements. The user should confirm the changes before they are processed.
 #'
-#' @param tdObject An object of class \code{treedata.table}
-#' @param taxa A vector class \code{character} containing all taxa that needs to be dropped from the original \code{treedata.table} object
-#' @return An object of class \code{treedata.table} with matching \code{$dat} and \code{$phy} elements based on whether \code{taxa} were dropped or not.
+#' @param tdObject An object of class `treedata.table`
+#' @param taxa A vector class `character` containing all taxa that needs to be dropped from the original `treedata.table` object
+#' @return An object of class `treedata.table` with matching `$dat` and `$phy` elements based on whether `taxa` were dropped or not.
 #' @examples
 #' data(anolis)
 #' # With a multiphylo object in the treedata.table object
@@ -21,7 +21,6 @@
 
 
 droptreedata.table <- function(tdObject, taxa) {
-
   if (!inherits(tdObject, c("treedata.table"))) {
     stop("Please use a class 'treedata.table' object \n")
   }
@@ -30,28 +29,33 @@ droptreedata.table <- function(tdObject, taxa) {
     stop("Please use a class 'character' object for taxa \n")
   }
 
-    .phy<-tdObject$phy
-    .dat<-tdObject$dat
+  .phy <- tdObject$phy
+  .dat <- tdObject$dat
 
-    if (inherits(.phy, c("phylo"))) {
-      .dat <- .dat[!.phy$tip.label %in% taxa]
-      .phy <- ape::drop.tip(.phy, which(.phy$tip.label %in% taxa))
-    } else {
-      .dat <- .dat[!.phy[[1]]$tip.label %in% taxa]
-      .phy <- lapply(.phy, ape::drop.tip, tip = which(.phy[[1]]$tip.label %in% taxa))
-      class(.phy) <- "multiPhylo"
-    }
+  if (inherits(.phy, c("phylo"))) {
+    .dat <- .dat[!.phy$tip.label %in% taxa]
+    .phy <- ape::drop.tip(.phy, which(.phy$tip.label %in% taxa))
+  } else {
+    .dat <- .dat[!.phy[[1]]$tip.label %in% taxa]
+    .phy <- lapply(.phy, ape::drop.tip, tip = which(.phy[[1]]$tip.label %in% taxa))
+    class(.phy) <- "multiPhylo"
+  }
 
 
-    tdObject$dat <- .dat
-    tdObject$phy <- .phy
-    attr(tdObject, "data_not_tree") <- if(attr(tdObject, "data_not_tree") =="OK"){
-      taxa}else{ c(attr(tdObject, "data_not_tree"),taxa)}
-    attr(tdObject, "tree_not_data") <- if(attr(tdObject, "tree_not_data") =="OK"){
-      taxa}else{ c(attr(tdObject, "tree_not_data"),taxa)}
-    attr(tdObject, "modified") <- 1
-    message(length(taxa), " taxa were dropped from the ORIGINAL treedata.table object")
+  tdObject$dat <- .dat
+  tdObject$phy <- .phy
[…]
