+++ R/detectCharacterType.R
@@ -104,13 +104,14 @@ forceNames <- function(dat, nameType = "row") {
   if (nType == "row" | nType == "rowcol") {
     if (!hasNames(dat, nameType = "row")) {
       nrows <- dim(dat)[1]
-      rownames(dat) <- paste("n", seq_along(dat[,1]), sep = "")
+      rownames(dat) <- paste("n", seq_along(dat[, 1]), sep = "")
     }
   }
   if (nType == "col" | nType == "rowcol") {
     if (!hasNames(dat, nameType = "col")) {
       ncols <- dim(dat)[2]
-      rownames(dat) <- paste("n", seq_along(dat[1,]), sep = "")     }
+      rownames(dat) <- paste("n", seq_along(dat[1, ]), sep = "")
+    }
   }
 
   dat
