+++ man/hasNames.Rd
@@ -11,10 +11,10 @@ hasNames(dat, nameType = "row")
 
 \item{nameType, }{either:
 \describe{
-    \item{"row"}{Rows}
-       \item{"col"}{Columns}
-         \item{"rowcol"}{Both rows and columns}
-    }}
+\item{"row"}{Rows}
+\item{"col"}{Columns}
+\item{"rowcol"}{Both rows and columns}
+}}
 }
 \description{
 Row and column name check

[…]
