### 2020-08-06 b5f19cb
REVIEW:name_column argument to explain that "detect"
