+++ R/as.treedata.table.R
@@ -10,7 +10,7 @@
 #' @importFrom data.table setDT as.data.table
 #' @param tree A tree of class `phylo` or multiple trees of class `multiPhylo`
 #' @param data A dataset in format `data.frame`
-#' @param name_column A character indicating the name of taxa in `data.frame`
+#' @param name_column A character indicating the name of taxa in `data.frame`. If set to `detect` (the default) `as treedata.table` will auto-detect this column
 #' @return treedata.table An object of type `treedata.table` containing the tree and data.table
 #' @importFrom ape drop.tip
 #' @importFrom geiger name.check
