### 2020-08-06 10354e2
REVIEW:remove @return in as.treedata.table()
