### 2020-08-06 2b62d02
REVIEW:Documentation for hasNames() and forceNames() updated
