### 2020-08-06 3acb5cc
REVIEW:multiple column names in extractVector
