+++ man/pulltreedata.table.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pull.treedata.table.R
-\name{pull.treedata.table}
-\alias{pull.treedata.table}
+% Please edit documentation in R/pulltreedata.table.R
+\name{pulltreedata.table}
+\alias{pulltreedata.table}
 \title{Returns the character matrix or phylogeny from a treedata.table object}
 \usage{
-pull.treedata.table(tdObject, type = "dat")
+pulltreedata.table(tdObject, type = "dat")
 }
 \arguments{
 \item{tdObject}{A treedata.table object}
@@ -20,6 +20,6 @@ Returns the character matrix or phylogeny from a treedata.table object
 \examples{
 data(anolis)
 td <- as.treedata.table(anolis$phy, anolis$dat)
-pull.treedata.table(td, type = "phy")
-pull.treedata.table(td, type = "dat")
+pulltreedata.table(td, type = "phy")
+pulltreedata.table(td, type = "dat")
 }
