+++ NAMESPACE
@@ -14,7 +14,7 @@ export(extractVector)
 export(filterMatrix)
 export(forceNames)
 export(hasNames)
-export(pull.treedata.table)
+export(pulltreedata.table)
 export(tdt)
 importFrom(ape,drop.tip)
 importFrom(data.table,as.data.table)
