### 2020-08-07 8fbaaa6
REVIEW:updated head and tail documentation
