+++ vignettes/treedata.table.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Getting Started With The treedata.table Package"
-author: "the vignette on the orient express"
+author: "Josef Uyeda, Cristian Roman-Palacios, April Wright"
 date: "12/22/2019"
 output: rmarkdown::html_vignette
 vignette: >
@@ -72,6 +72,8 @@ Multiple traits can also be extracted using `extractVector`.
 extractVector(td, 'SVL','ecomorph')
 ```
 
+However, there's a couple aspects that are unique to `[[.treedata.table()` and `extractVector()`. First, `[[.treedata.table()` has an extra exact argument to enable partial match. Second, `extractVector()` can extract multiple columns and accepts non-standard evaluation.
+
 The real power in treedata.table is in co-indexing the tree and table. For example, in the below command, we use `data.table` syntax to take the first representative from each ecomorph. We retain all data columns. If you examine the tree object, you will see that it has had all the tips not in the resultant `data.table` dropped.
 
 ```{r}
@@ -109,31 +111,28 @@ tdt(td, geiger::fitContinuous(phy, extractVector(td, 'SVL'), model="BM", ncores=
 ```
 
 
-We can also drop tips directly from the tree, and have those tips drop concurrently from the data.table. In the example below, we  remove two taxa by name. Please note that the below is intended as an interactive function.
+We can also drop tips directly from the tree, and have those tips drop concurrently from the data.table. In the example below, we  remove two taxa by name. 
 
-```{r, eval=FALSE}
-to_drop <- c("chamaeleonides" ,"eugenegrahami" )
-
-dt <- droptreedata.table(tdObject=td, taxa=to_drop)
-
-to_drop %in% dt$phy$tip.label
-
-[1] FALSE FALSE
-
-to_drop %in% dt$dat$X
-
-[1] FALSE FALSE
-
-
+```{r}
+dt <- droptreedata.table(tdObject=td, taxa=c("chamaeleonides" ,"eugenegrahami" ))
 ```
 
+We can check if *A. chamaeleonides* and *A. eugenegrahami* are still in the tree
 
+```{r}
+c("chamaeleonides" ,"eugenegrahami" ) %in% dt$phy$tip.label
+```
+
+And we can do the same with the data in out `treedata.table` object
+
+```{r}
+c("chamaeleonides" ,"eugenegrahami" ) %in% dt$dat$X
+```
 
 
 When you're done, the data.table and tree can both be extracted from the object:
 
 ```{r}
-
 df <- pulltreedata.table(td, "dat")
 tree <- pulltreedata.table(td, "tree")
 ```
