### 2020-08-07 abfb2cb
reconfigure author block so ORCIDs can be added
