### 2020-08-07 9cd075a
updated README per review
