### 2020-08-07 a9a7bde
added CONTRIBUTING document
