+++ tests/testthat/test.treedata.table.R
@@ -29,7 +29,8 @@ test_that("Species in tree and trait dataframe are the same", {
 
 
 test_that("datasets are the same", {
-  expect_true(all(colnames(anolis$dat)[-1] == colnames(as.data.frame(td$dat))[-1]))
+  expect_true(all(colnames(anolis$dat)[-1] ==
+    colnames(as.data.frame(td$dat))[-1]))
 })
 
 
@@ -66,7 +67,8 @@ test_that("[[ extracts a named character vector", {
 
 
 
-test_that("phy and dat objects can be extracted correctly using pulltreedata.table", {
+test_that("phy and dat objects can be extracted correctly using
+          pulltreedata.table", {
   expect_is(pulltreedata.table(td, type = "phy"), "phylo")
   expect_is(pulltreedata.table(td, type = "dat"), "data.table")
 })
@@ -80,8 +82,12 @@ test_that("Column containing tip labs can be correctly detected", {
   expect_equal(td1$phy, td2$phy)
 })
 
-test_that("Find the correct number of discrete/continuous characters in the anolis dataset", {
-  expect_equal(detectCharacterType(anolis$dat[, 1]), detectAllCharacters(anolis$dat)[1])
+test_that("Find the correct number of discrete/continuous characters in the
+          anolis dataset", {
+  expect_equal(
+    detectCharacterType(anolis$dat[, 1]),
+    detectAllCharacters(anolis$dat)[1]
+  )
 })
 
 
@@ -100,7 +106,8 @@ test_that("Error is shown when tips with different tip labels are used", {
   tree2 <- list(anolis$phy, anolis2)
   class(tree2) <- "multiPhylo"
   expect_error(as.treedata.table(tree = tree2, data = as.anolis$dat),
-    "Please make sure that tip labels are equivalent across trees in the multiPhylo object \n",
+    "Please make sure that tip labels are equivalent across trees in the
+    multiPhylo object \n",
     fixed = T
   )
 })
@@ -115,9 +122,12 @@ test_that("Error is a non-phylo (or multiPhylo) object is used in the phy", {
 
 
 test_that("Error is a non-data.frame is used in as.treedata.table", {
-  expect_error(as.treedata.table(tree = anolis$phy, data = as.matrix(anolis$dat)),
-    "Your data MUST be of class data.frame",
-    fixed = T
+  expect_error(as.treedata.table(
+    tree = anolis$phy,
+    data = as.matrix(anolis$dat)
+  ),
+  "Your data MUST be of class data.frame",
+  fixed = T
   )
 })
 
@@ -158,7 +168,8 @@ test_that("Normal as.treedata.table with data.frame without row.names
   )
 })
 
-test_that("Normal as.treedata.table but testing if the tips dropped message is shown for trees dropped from tree", {
+test_that("Normal as.treedata.table but testing if the tips dropped message is
+          shown for trees dropped from tree", {
   anolis1 <- anolis$phy
   anolis1$tip.label[1] <- "NAA"
 
@@ -169,7 +180,8 @@ test_that("Normal as.treedata.table but testing if the tips dropped message is s
 })
 
 
-test_that("Normal as.treedata.table but testing if the tips dropped message is shown for trees dropped from data", {
+test_that("Normal as.treedata.table but testing if the tips dropped message is
+          shown for trees dropped from data", {
   anolis1 <- anolis$phy
   anolis1$tip.label[1] <- "NAA"
 
@@ -180,36 +192,51 @@ test_that("Normal as.treedata.table but testing if the tips dropped message is s
 })
 
[…]
