### 2020-08-07 d860070
REVIEW: make tests compatible with 80-line limits
