+++ R/detectCharacterType.R
@@ -15,8 +15,7 @@ detectCharacterType <- function(dat, repeatsAsDiscrete = TRUE, cutoff = 0.1) {
   if (is.factor(dat)) {
     charType <- "discrete"
   } else if (nlevels(as.factor(dat)) / length(dat) < cutoff) {
-    warning("Guessing that this is a discrete character based on repeated
-              values")
+    warning("Guessing this is a discrete character based on repeated values")
     charType <- "discrete"
   } else {
     charType <- "continuous"
