### 2020-08-08 3253251
REVIEW:Removing charType from filterMatrix
