+++ man/detectCharacterType.Rd
@@ -9,9 +9,12 @@ detectCharacterType(dat, repeatsAsDiscrete = TRUE, cutoff = 0.1)
 \arguments{
 \item{dat}{A vector of data}
 
-\item{repeatsAsDiscrete}{If TRUE, consider numeric variables that repeat values exactly as discrete; see cutoff}
+\item{repeatsAsDiscrete}{If TRUE, consider numeric variables that repeat
+values exactly as discrete; see cutoff}
 
-\item{cutoff}{Cutoff value for deciding if numeric data might actually be discrete: if nlev is the number of levels and n the length of dat, then nlev / n should exceed cutoff, or the data will be classified as discrete}
+\item{cutoff}{Cutoff value for deciding if numeric data might actually be
+discrete: if nlev is the number of levels and n the length of dat, then
+nlev / n should exceed cutoff, or the data will be classified as discrete}
 }
 \value{
 Either "discrete" or "continuous"
