+++ man/as.treedata.table.Rd
@@ -2,7 +2,8 @@
 % Please edit documentation in R/as.treedata.table.R
 \name{as.treedata.table}
 \alias{as.treedata.table}
-\title{Combine tree (or set of trees) and data.frame into a single treedata.table object}
+\title{Combine tree (or set of trees) and data.frame into a single treedata.table
+object}
 \usage{
 as.treedata.table(tree, data, name_column = "detect")
 }
@@ -11,15 +12,18 @@ as.treedata.table(tree, data, name_column = "detect")
 
 \item{data}{A dataset in format \code{data.frame}}
 
-\item{name_column}{A character indicating the name of taxa in \code{data.frame}. If set to \code{detect} (the default) \verb{as treedata.table} will auto-detect this column}
+\item{name_column}{A character indicating the name of taxa in \code{data.frame}.
+If set to \code{detect} (the default) \verb{as treedata.table} will auto-detect this
+column}
 }
 \value{
 An object of type \code{treedata.table} containing the tree and data.table
 }
 \description{
-This function takes as input a tree of class \code{phylo} or \code{multiPhylo} and a \code{data.frame}
-and combines them into a treedata.table. If a \code{multiPhylo} is provided, all trees must have the same tip.labels.
-\code{treedata.table} object is sorted such that the rows in the data.table are matched to the tip.labels
+This function takes as input a tree of class \code{phylo} or \code{multiPhylo} and a
+\code{data.frame} and combines them into a treedata.table. If a \code{multiPhylo} is
+provided, all trees must have the same tip.labels. \code{treedata.table} object is
+sorted such that the rows in the data.table are matched to the tip.labels
 of the phylogeny. Tip.labels on the tree must match a column of tip
 names in the input data.frame. The output of this function will be a
 treedata.table, which can be manipulated as a data.table.
