+++ man/forceNames.Rd
@@ -20,8 +20,8 @@ forceNames(dat, nameType = "row")
 An object of type `data.frame with labeled columns, rows, or both.
 }
 \description{
-This function creates column names (\code{colnames}), row.names (\code{row.names}), or both
-in an unnamed \code{data.frame} or \code{matrix}.
+This function creates column names (\code{colnames}), row.names (\code{row.names}),
+or both in an unnamed \code{data.frame} or \code{matrix}.
 }
 \examples{
 data(anolis)

[…]
