### 2020-08-08 ca94642
REVIEW:Description added to detectCharacterType, detectCharacterChanges, and filterMatrix
