### 2020-08-08 9988f78
REVIEW:Vignette updated for filterMatrix, forceNames, and hasNames
