+++ R/detectCharacterType.R
@@ -5,8 +5,6 @@
 #' (e.g., with values blue, red, yellow) character.
 #'
 #' @param dat A vector of data
-#' @param repeatsAsDiscrete If TRUE, consider numeric variables that repeat
-#' values exactly as discrete; see cutoff
 #' @param cutoff Cutoff value for deciding if numeric data might actually be
 #' discrete: if nlev is the number of levels and n the length of dat, then
 #' nlev / n should exceed cutoff, or the data will be classified as discrete
@@ -15,7 +13,7 @@
 #' data(anolis)
 #' detectCharacterType(anolis$dat[, 1])
 #' @export
-detectCharacterType <- function(dat, repeatsAsDiscrete = TRUE, cutoff = 0.1) {
+detectCharacterType <- function(dat, cutoff = 0.1) {
   if (is.factor(dat)) {
     charType <- "discrete"
   } else if (nlevels(as.factor(dat)) / length(dat) < cutoff) {
@@ -34,8 +32,6 @@ detectCharacterType <- function(dat, repeatsAsDiscrete = TRUE, cutoff = 0.1) {
 #' (e.g., with values blue, red, yellow).
 #'
 #' @param mat A matrix of data
-#' @param repeatsAsDiscrete If TRUE, consider numeric variables that repeat
-#' values exactly as discrete; see cutoff
 #' @param cutoff Cutoff value for deciding if numeric data might actually be
 #' discrete: if nlev is the number of levels and n the length of dat, then
 #' nlev / n should exceed cutoff, or the data will be classified as discrete
@@ -45,11 +41,11 @@ detectCharacterType <- function(dat, repeatsAsDiscrete = TRUE, cutoff = 0.1) {
 #' data(anolis)
 #' detectAllCharacters(anolis$dat)
 #' @export
-detectAllCharacters <- function(mat, repeatsAsDiscrete = TRUE, cutoff = 0.1) {
+detectAllCharacters <- function(mat, cutoff = 0.1) {
   nchar <- dim(mat)[2]
   result <- numeric(nchar)
   for (i in seq_len(nchar)) {
-    result[i] <- detectCharacterType(mat[, i], repeatsAsDiscrete, cutoff)
+    result[i] <- detectCharacterType(mat[, i], cutoff)
   }
   return(result)
 }
