+++ man/detectCharacterType.Rd
@@ -4,13 +4,10 @@
 \alias{detectCharacterType}
 \title{Function to detect whether a character is continuous or discrete}
 \usage{
-detectCharacterType(dat, repeatsAsDiscrete = TRUE, cutoff = 0.1)
+detectCharacterType(dat, cutoff = 0.1)
 }
 \arguments{
 \item{dat}{A vector of data}
-
-\item{repeatsAsDiscrete}{If TRUE, consider numeric variables that repeat
-values exactly as discrete; see cutoff}
 
 \item{cutoff}{Cutoff value for deciding if numeric data might actually be
 discrete: if nlev is the number of levels and n the length of dat, then
