+++ man/detectAllCharacters.Rd
@@ -4,13 +4,10 @@
 \alias{detectAllCharacters}
 \title{Apply detectCharacterType over an entire matrix}
 \usage{
-detectAllCharacters(mat, repeatsAsDiscrete = TRUE, cutoff = 0.1)
+detectAllCharacters(mat, cutoff = 0.1)
 }
 \arguments{
 \item{mat}{A matrix of data}
-
-\item{repeatsAsDiscrete}{If TRUE, consider numeric variables that repeat
-values exactly as discrete; see cutoff}
 
 \item{cutoff}{Cutoff value for deciding if numeric data might actually be
 discrete: if nlev is the number of levels and n the length of dat, then
