### 2020-08-09 42b58cf
Add logo
