### 2020-08-09 2a12a7d
Logo size
