### 2020-08-09 7062a02
Logo resized
