+++ man/anolis.Rd
@@ -4,9 +4,7 @@
 \name{anolis}
 \alias{anolis}
 \title{Anole data}
-\format{
-An object of class \code{list} of length 2.
-}
+\format{An object of class \code{list} of length 2.}
 \usage{
 data(anolis)
 }
