+++ DESCRIPTION
@@ -5,20 +5,26 @@ Authors@R: c(
     person(given = "Josef",
            family = "Uyeda", 
            role=c("aut"),
-           email="juyeda@vt.edu", 
-          comment = c(ORCID="0000-0003-4624-9680")
+           email="juyeda@vt.edu",
+           comment = c(ORCID="0000-0003-4624-9680")
            ),
     person(given = "Cristian", 
             family = "Roman-Palacios",
             role = c("cre"),
             email = "cromanpa94@email.arizona.edu", 
-          comment = c(ORCID="0000-0003-1696-4886")
+          comment = c(ORCID="0000-0003-4692-3225")
             ),
     person(given = "April",
           family = "Wright",
           role = c("aut"),
           email = "april.wright@southeastern.edu", 
-          comment = c(ORCID="0000-0003-4692-3225"))
+          comment = c(ORCID="0000-0003-4692-3225")),
+    person(given = "Hugo",
+          family = "Gruson",
+          role = c("rev")),
+    person(given = "Kari",
+          family = "Norman",
+          role = c("rev"))
   )
 Maintainer: Cristian Roman-Palacios <cromanpa94@email.arizona.edu>
 URL: https://github.com/uyedaj/treedata.table/
