+++ NAMESPACE
@@ -18,7 +18,6 @@ export(pulltreedata.table)
 export(tdt)
 importFrom(ape,drop.tip)
 importFrom(data.table,as.data.table)
-importFrom(data.table,setDT)
 importFrom(geiger,name.check)
 importFrom(lazyeval,lazy_dots)
 importFrom(utils,head)
