+++ R/treedata.table.R
@@ -55,25 +55,26 @@
 
 `[.treedata.table` <- function(x, ...) {
   .dat <- x$dat
-  .dat <- base::cbind(.dat, "rowid" = seq_len(nrow(.dat))) # CRP: using cbind
+  .dat <- base::cbind(.dat, "rowid" = seq_len(nrow(.dat)))
   # instead of :=
   dots <- lazyeval::lazy_dots(...)
   if ("by" %in% names(dots)) {
     .dat <- .dat[...]
   } else {
     if (nchar(dots[[2]]$expr)[1] != 0) {
-      .dat <- .dat[..., by = "rowid"] # CRP: using "rowid" instead of rowid
+      .dat <- .dat[..., by = "rowid"]
     } else {
       .dat <- .dat[...]
     }
   }
+
   # .phy <- ape::drop.tip(x$phy, which(!1:nrow(x$dat) %in%
   #  unlist(.dat[, "rowid"]))) #CRP: using "rowid" instead of rowid & unlist
 
   .phy <- if (inherits(x$phy, c("phylo"))) {
-    ape::drop.tip(x$phy, which(!seq_along(x$dat) %in% unlist(.dat[, "rowid"])))
+    ape::drop.tip(x$phy, which(!1:nrow(x$dat) %in% unlist(.dat[, "rowid"])))
   } else {
-    tr <- lapply(x$phy, ape::drop.tip, tip = which(!seq_along(x$dat)
+    tr <- lapply(x$phy, ape::drop.tip, tip = which(!1:nrow(x$dat)
     %in% unlist(.dat[, "rowid"])))
     class(tr) <- "multiPhylo"
     tr
