### 2020-08-19 uyedaj/treedata.table#1 @cromanpa94
Reviews at rOpenSci
# Response to reviewers - treedata.table

### About

In this document, we address all the comments raised during the review phase of `treedata.table` in rOpenSci. When possible, we use a single commit to answer each of the comments.

### Where are changes implemented?

All the changes listed below are implemented in the following branch of our GitHub repository:

https://github.com/uyedaj/treedata.table/tree/cristian

### Acknowledgements

We thank two @Bisaloo (`reviewer`), @karinorman (`reviewer`), @jooolia (`editor`) for their comments!
___

#### First reviewer (@Bisaloo):

- [x] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("`rev`" role) in the package `DESCRIPTION` file.

https://github.com/uyedaj/treedata.table/commit/44ccedab95d9df3109815f5dbbeddf1ee608a962

- [x] It's recommended practice to add an `ORCiD` for the authors if they have one.

https://github.com/uyedaj/treedata.table/commit/abfb2cb95554c66888bd9b3f03e338f741018626
https://github.com/uyedaj/treedata.table/commit/b76f4a8fdc38160ef6df8e40d16bdb05d2a907e4

- [x] Please add a contributing guide.

https://github.com/uyedaj/treedata.table/commit/a9a7bde532e4f625e8cace4d38f139edaf0c09c3
https://github.com/uyedaj/treedata.table/commit/16b6fbe767dec1a154cff21272bb34e75faa6aff

- [x] As far as I know, the Type field in the `DESCRIPTION` is not standard and in all cases unnecessary

https://github.com/uyedaj/treedata.table/commit/07746cffbd84d697e5ad1c5d424a6eb6677c3001

- [x] The package `Title` is not actually using title case. I don't care much personally but in my experience, this can cause issues when you submit to CRAN. You can verify this with the `tools::toTitleCase()` function.

https://github.com/uyedaj/treedata.table/commit/ff610ad869f79b5c433937af27bac40491853259

- [x] Please avoid as much as possible to use `Depends` when you can use `Imports`. Here is a relevant quote from Hadley Wickham and Jenny Bryan's book on R package development: Depends: Prior to the rollout of namespaces in R 2.14.0, Depends was the only way to “depend” on another package. Now, despite the name, you should almost always use Imports, not Depends. You’ll learn why, and when you should still use Depends, in namespaces.

https://github.com/uyedaj/treedata.table/commit/3d43122f7d52a7fd415fd05ecaabe5fe4951bb31

- [x] According to the CRAN guidelines for MIT licenses, your previous version of `LICENSE` was correct (before `uyedaj/treedata.table@7764ef2`). You just need to update to name of the copyright holder(s).

https://github.com/uyedaj/treedata.table/commit/8b6250e850db5982686bfcec468af06a14b67bbc

- [x] The `README` does not contain enough information about the package, how to install and use it. You may find this relevant chapter of rOpenSci's devguide useful.

https://github.com/uyedaj/treedata.table/commit/9cd075a01b3fcf7f1d6f9e2baccb6a402f8e71c9

- [x] Since this package is presented as addressing a performance need, it would be very nice to see some benchmarks to back this up. You can use the microbenchmark package for this.

https://github.com/uyedaj/treedata.table/commit/3bd657411bfa8f709ac597fe754a1b7d016c9abb

- [x] (optional) it can be nice to add more badges to your `README` to advertise the fact that you follow the current best practices in package development. In particular, I like badges with code coverage. Here are the instructions on how to do this with `codecov`.

https://github.com/uyedaj/treedata.table/commit/b38d88ad0a6fa94f38effe30a789ba81c0846a9c

- [x] Please avoid multiline instructions (separated by `;`). It makes the code harder the read. This will also help with the long lines warning reported by goodpratice in @jooolia's comment:

https://github.com/uyedaj/treedata.table/commit/59872e58f1bb21d92acd8b7f2a3f5296fb25f3ad
https://github.com/uyedaj/treedata.table/commit/748ddde2b5c9e3843b89b6935abd737fd3ac7622

- [x] When you perform tests on objects that are already booleans (`TRUE`/`FALSE`), you don't need to `write == TRUE`.

https://github.com/uyedaj/treedata.table/commit/9afe310bdb05b68fb6518f0ca919b48f26a35c3b

- [x] Class testing should be done with `inherits()` (for `S3` objects, which is the case for ape objects), or `is()` (for `S4` objects). Please see this blog post by Martin Maechler for more info.

```
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
p <- pull.treedata.table(td, type = "phy")
d <- pull.treedata.table(td, type = "dat")
td2 <- as.treedata.table(p, d)
# You would expect to get identical(td, td2) but this line actually throws an error
```

`as.treedata.table` uses a data.frame as input:

```
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
p <- pull.treedata.table(td, type = "phy")
d <- pull.treedata.table(td, type = "dat")
td2 <- as.treedata.table(p, as.data.frame(d))
> identical(td,td2)
[1] TRUE
```

https://github.com/uyedaj/treedata.table/commit/1b5c188f15b4e85eeb5ca668d7e8ae8375ad738a

- [x] Regarding your replacement of `1:length()` by `seq_len()` in `uyedaj/treedata.table@7764ef2`, the correct syntax of `seq_len()` doesn't use `1`:.

https://github.com/uyedaj/treedata.table/commit/700e0edf256c2e31d5a9b5d57bf1d6d8c0413f7e
https://github.com/uyedaj/treedata.table/commit/19a777c9c12bf8320cb34533cc0a508ff6c16a06
https://github.com/uyedaj/treedata.table/commit/977959294d3cc30cad8c5127f9a8eb38b3349502
https://github.com/uyedaj/treedata.table/commit/96f84da8e01fb24756b5b2abd93e53b518b193b3

- [x] Communicate information to users via the `message()` function instead of `cat()`. This will always surely cause issues upon CRAN submission if not corrected. See this detailed SO answer for more info about the differences between `message()` and `cat()`.

https://github.com/uyedaj/treedata.table/commit/b99ea3c1d69259cf4800350c0a0fcd6302c7a6c0

- [x] On a related note, you don't need to use `paste0()` in `message()` calls. The `message()` will automatically concatenate the arguments you give it.

https://github.com/uyedaj/treedata.table/commit/c7ebb3b3026cd9dbb37b816c03761ad730fe5969

- [x] There is a problem in `as.treedata.table()` with the name_column argument not working. No matter what the user enters in name_column or what the auto-detection code finds, you always use the first column for `tip.labels`.

https://github.com/uyedaj/treedata.table/commit/cb04cf56106a3835259c8fbe14eaea40d45a06f2

- [x] This behaviour should be tested in tests.

https://github.com/uyedaj/treedata.table/commit/e00fef93a65854807824fad61b3e6dae73587504

- [x] (optional) it would be nice to have a message indicating which column was auto-detected as containing the tip.labels when `name_column = "detect"` in `as.treedata.table()`

https://github.com/uyedaj/treedata.table/commit/5b1951b8c4c3c9ee261d3b4052ef3e80101b00a1

- [x] (optional) uniformise the indents in your code. You sometimes use tabs and sometimes spaces. This can be fixed by running `styler::style_pkg()` in the root of your RStudio project.

https://github.com/uyedaj/treedata.table/commit/d8600705b83e9dc7cb64d06d19fce0262436016b
https://github.com/uyedaj/treedata.table/commit/d4cffa588bedc16bb3822164f29babe7909350f0
https://github.com/uyedaj/treedata.table/commit/cd0c945561f9b110cc598f9f21b08e95675aae75
https://github.com/uyedaj/treedata.table/commit/016bc066cf0c205abc665e64d55c2cfdf6cdf35a
https://github.com/uyedaj/treedata.table/commit/bfb25723f75658d4d9b39f706183f9a148228940

- [x] As far as I can tell, the repeatsAsDiscrete is not used in `detectCharacterType()` or `detectAllCharacters()`

https://github.com/uyedaj/treedata.table/commit/79cfcb4f6231a7a812f622848a4309d2581d34bd

- [x] This line will cause wrong identifications since the conversion of a data.frame to a matrix forces all elements to the same type.

https://github.com/uyedaj/treedata.table/commit/8d4b9e1920e81cc3f172552105d096748f6e8dd4

- [x] Regarding my later point about tests, this would make a good test: ensuring that you find the correct number of `discrete`/`continuous` characters in the anolis dataset.

https://github.com/uyedaj/treedata.table/commit/3cd17b95124a4538c93200336dbc1d5df864a62a
https://github.com/uyedaj/treedata.table/commit/4752e44664a56613e54b5befe6be574890f231a9

- [x] I think you missed some `1:` instances from @jooolia's comment. E.g.,

https://github.com/uyedaj/treedata.table/commit/19a777c9c12bf8320cb34533cc0a508ff6c16a06
https://github.com/uyedaj/treedata.table/commit/977959294d3cc30cad8c5127f9a8eb38b3349502
https://github.com/uyedaj/treedata.table/commit/96f84da8e01fb24756b5b2abd93e53b518b193b3

- [x] It may be my unfamiliarity with data.table but I don't understand the warning/prompt in `droptreedata.table()`. As far as I can tell, the original data is NOT modified. Using your example:

```
data(anolis)
td <- as.treedata.table(anolis$phy, anolis$dat)
td_old <- td
td_new <- droptreedata.table(tdObject = td, taxa = c("chamaeleonides" ,"eugenegrahami"))

identical(td, td_old)
#> TRUE

identical(td, td_new)
#> FALSE
```

`droptreedata.table()` is used to remove species from a `treedata.table` object. In your example, 2 species (*chamaeleonides* and *eugenegrahami*) are being excluded from the `treedata.table` object:
```
> data(anolis)
> td <- as.treedata.table(anolis$phy, anolis$dat)
Tip labels detected in column: X
Phylo object detected

No tips were dropped from the original tree/dataset
> td_old <- td
> td_new <- droptreedata.table(tdObject = td, taxa = c("chamaeleonides" ,"eugenegrahami"))
Please confirm that you would like to make changes to the ORIGINAL data?
Type: (1) YES, (2) NO: 1
> nrow(td_new$dat)
[1] 98
> nrow(td_old$dat)
[1] 100
```

The message was used to inform the user that n taxa are being dropped from the dataset. **We have removed this prompt from the revised version**. Regarding your tests, (1)`identical(td, td_old)` must be `TRUE` because these two are duplicates (`td_old <- td`). However, (2) `identical(td, td_new)` must be `FALSE` given two taxa were dropped from the latter object (please see the difference in row numbers between both objects)

https://github.com/uyedaj/treedata.table/commit/a6d0b6048efcf2b1cc51b224c2f6026338ed55f4

- [x] The dots handling in `extractVector()` may probably be simplified with `lazyeval` since you already have it as a dependency

https://github.com/uyedaj/treedata.table/commit/63a2935c34af83ff020042bf4cb6bb424720ad54

- [x] It would be useful to add a `match.arg(type, c("dat", "phy"))` in pull.treedata.table()

https://github.com/uyedaj/treedata.table/commit/9408ac3e13497f815806cc3d7608f42c94bab518

- [ ] In `tdt`, do you really need `...`? Wouldn't an `FUN` arg be sufficient?

Although we fully agree with the reviewer, both approaches are essentially doing the same. We decided to keep the ellipsis.

- [x] Since you provide a head() method for `treedata.table`, it would be nice to have a `tail()` method as well

https://github.com/uyedaj/treedata.table/commit/8fbaaa6da774b496fcced6bf789cba4cd4a95c69
https://github.com/uyedaj/treedata.table/commit/9b62811cbe7ff52c1d76eb3092232ecd61066885

- [x] `paste()`s are unnecessary here since `cat()`

https://github.com/uyedaj/treedata.table/commit/c7ebb3b3026cd9dbb37b816c03761ad730fe5969

- [x] The output of `summary.treedata.table()` is slightly confusing in my opinion. On the example you provide (with anolis), where no taxa are dropped from the tree or the data, you get:

We changed the message to `“Taxa dropped from the tree/data”`:
```
> data(anolis)
> td <- as.treedata.table(anolis$phy, anolis$dat)
Tip labels detected in column: X
Phylo object detected
No tips were dropped from the original tree/dataset
> summary(td)
A treedata.table object
The dataset contains 11 traits
Continuous traits: tip.label, SVL, PCI_limbs, PCII_head, PCIII_padwidth_vs_tail, PCIV_lamella_num, awesomeness, hostility, attitude
Discrete traits: ecomorph, island
The following traits have missing values: 0
Taxa dropped from the tree:  0
Taxa dropped from the data:  0
```

We also modified the `droptreedata.table()` function to keep track of the dropped species in the new object:

```
> td_new <- droptreedata.table(tdObject = td, taxa = c("chamaeleonides" ,"eugenegrahami"))
2 taxa were dropped from the ORIGINAL treedata.table object
> summary(td_new)
A treedata.table object
The dataset contains 11 traits
Continuous traits: tip.label, SVL, PCI_limbs, PCII_head, PCIII_padwidth_vs_tail, PCIV_lamella_num, awesomeness, hostility, attitude
Discrete traits: ecomorph, island
The following traits have missing values: 0
Taxa dropped from the tree: chamaeleonides, eugenegrahami
Taxa dropped from the data: chamaeleonides, eugenegrahami
```

https://github.com/uyedaj/treedata.table/commit/292295cf369975ffdac2b677772020bc96f8103f#diff-262ada0e7c14967a2a41b70149c367fb

- [x] When I run examples from `[.treedata.table()`, I get the following warning:

```
In 1:seq_along(x$dat) :
numerical expression has 11 elements: only the first used
```

We cannot replicate this warning in the latest version of the package. It was probably fixed in a previous commit!

- [ ] (optional) your tests are not compatible with the upcoming v3 of the testthat package. In particular, you could replace the (soon to be) deprecated `expect_is()` function by `expect_s3_class()`. Excepted these two lines

We will keep the current functions for now but we thank the reviewer for this comment.

- [x] I think the test coverage needs to be increased a lot, as shown by the bugs uncovered during this review. I sometimes noted what would be good candidates for tests. As a good starting point, you can also look at which lines are not covered by your tests on `codecov`: https://codecov.io/github/uyedaj/treedata.table?branch=master. I see no technical limitations that would prevent you to reach 100% coverage for this package but for now, I think you should at least try to reach 80% coverage.

The current version has >80% of coverage

https://github.com/uyedaj/treedata.table/commit/ab53e267dfa7c51f9615d122834f58716aca1dea

- [x] (optional) you may want to use markdown syntax in your `Roxygen` comments. This produces more readable documentation in the source file in my opinion. Automatic conversion of your current `Roxygen` comments to markdown can be done with the `roxygen2md` package.
Please add a sentence for the name_column argument to explain that "`detect`" (the default) will auto-detect this column:

https://github.com/uyedaj/treedata.table/commit/b5f19cbb2387c6e1c591e35b565688a47b663143

- [x] In `as.treedata.table()`, you declare but this is unnecessary since you don't use `setDT()` here and `as.data.table()` is prefixed with the `data.table::` namespace

https://github.com/uyedaj/treedata.table/commit/92bf3c887226f2c3f0f3d4c69e1f7f94ceacf7a8

- [x] I'm not sure what the first word means in your @return `roxygen` comments. I think it will be less confusing if you remove these(?)

https://github.com/uyedaj/treedata.table/commit/10354e2b258b42b7ca2ca55c1178e8f081160cd8

- [x] (optional) you can use the @inheritParams `detectCharacterType` `roxygen` comment in the documentation of `detectAllCharacters()` to avoid duplicating the documentation of the function arguments. This is useful because it ensures the documentation of these functions will always stay in sync in the future. No risk of forgetting to update one of the two!

We chose to leave all the parameters in both of the functions. We thank the reviewer for his comment.

- [x] (optional) I find it useful to explicitly state which is default for all arguments (when it exists) in the documentation. E.g., I would change it to something like `#'` @param returnType Either discrete (the default) or continuous

https://github.com/uyedaj/treedata.table/commit/fa8fd8f79553dd2025e78407c875fe3b5db11992

- [x] Please expand the documentation of the `hasNames()`/`forceNames()` functions by providing a short explanation of the function purpose and/or a @return `roxygen` comment.

https://github.com/uyedaj/treedata.table/commit/2b62d02718cffd9ae398b5c8026accbe2cf58eed
https://github.com/uyedaj/treedata.table/commit/3acb5cc58e1eea644d3a60d1792699ce29f2fc7f

- [x] The documentation fo `extractVector()` should be updated to indicate that multiple column names can be passed to `...`, which means you don't necessarily get a named vector but a list of named vectors, as opposed to what the documentation says.

https://github.com/uyedaj/treedata.table/commit/3acb5cc58e1eea644d3a60d1792699ce29f2fc7f
https://github.com/uyedaj/treedata.table/commit/0f31ded499064e5a2ed8caaf8cc4d483b364fdcf
https://github.com/uyedaj/treedata.table/commit/42491d2acf79c06ca1d4069849682f721f08055c

 - [x] I think the function name `pull.treedata.table` is slightly confusing as it sounds like a `S3` and it's not. Something along the lines of `pull_treedata.table()` or `pulldata.data.table()` (you already have a `dropdatatree.table()` function so this would make sense) or `???` would probably be better.

https://github.com/uyedaj/treedata.table/commit/cd8fc93b413c7a51f0f524331560804234ae1bfb

- [x] Shouldn't this just be "`If negative, all but the n last rows of x`" (i.e., remove "first"):

We have completely modified our previous `head()` function.

https://github.com/uyedaj/treedata.table/commit/500648a62fef5761eab22554773cafcadcf4f1a3

- [x] This doesn't seem correct. The dots are ignored in `head.treedata.table()` as far as I can tell

We have completely modified our previous `head()` function.

https://github.com/uyedaj/treedata.table/commit/500648a62fef5761eab22554773cafcadcf4f1a3

- [x] To fix the `NOTE` in R CMD check, you need to `importFrom(utils,head)` before you try to define another method for this generic. This can be done by adding a #' @importFrom utils head `roxygen` comment in the `roxygen` chunk of `head.treedata.table()` for example.

https://github.com/uyedaj/treedata.table/commit/500648a62fef5761eab22554773cafcadcf4f1a3

- [x] Please update the list of authors in the vignette

https://github.com/uyedaj/treedata.table/commit/07d2178cab836e717f9b48758f79c57ee8a9949c

- [x] (optional) it may useful for users to mention the differences between `[[.treedata.table() `and `extractVector()`, namely that `[[.treedata.table()` has an extra exact argument to enable partial match while extractVector() can extract multiple columns and accepts non-standard evaluation

https://github.com/uyedaj/treedata.table/commit/07d2178cab836e717f9b48758f79c57ee8a9949c

- [x] If you remove the "`[1] FALSE FALSE`" (which are actually output, and not R code) from this chunk, you will be able to remove the `eval = FALSE`. It's awlays better when all chunks run and can be copied/pasted in the console directly

https://github.com/uyedaj/treedata.table/commit/07d2178cab836e717f9b48758f79c57ee8a9949c

___
### Second Reviewer (@karinorman):

- [x] The warnings for co-indexing calls (e.g. `td[, head(.SD, 1), by = "ecomorph"]` from the vignette don't seem particularly informative, I would suppress unless necessary.

We cannot replicate this warning in the latest version of the package. It seems to be fixed now!

- [x] I think having the user confirm the changes for `droptreedata.table` is unnecessary. For me at least it made me assume that the function was modifying an object in place, which I realized it wasn't after some exploration.

https://github.com/uyedaj/treedata.table/commit/a6d0b6048efcf2b1cc51b224c2f6026338ed55f4

- [ ] Consider changing the naming convention to *.td.table rather than *.`treedata.table` for the sake of brevity. You could also add a td to other exported functions (e.g. td.`extractVector`).

We thank the reviewer very much but we have decided to keep the current names.

[…]
