### 2020-08-19 #367 @cromanpa94
Thank you so much @karinorman and @Bisaloo for your comments 😄

We have addressed your concerns in a new branch (https://github.com/uyedaj/treedata.table/tree/cristian). Our answers are presented under issues in our repository (https://github.com/uyedaj/treedata.table/issues/1) but we're more than happy to post the same information here if necessary.
