### 2020-08-20 d2ba686
Restore cat() is print.treedata.table() cat() is appropriate here (this is the only case) since the purpose of print() functions is to display something. See for example the base function print.data.frame()
