### 2020-08-20 c5ecb2f
Fix extra whitespace in message()
