+++ R/as.treedata.table.R
@@ -131,8 +131,8 @@ as.treedata.table <- function(tree, data, name_column = "detect") {
       data <- data[!as.character(data[, name_column]) == data_not_tree, ]
       message(
         length(c(tree_not_data)), " tip(s) dropped from ", length(tree),
-        " trees", "\n", length(c(data_not_tree)), " tip(s)  dropped from the
-                                                    original dataset"
+        " trees", "\n", length(c(data_not_tree)), " tip(s)  dropped from the ",
+        " original dataset"
       )
     } else {
       message("No tips were dropped from the original trees/dataset")
