### 2020-08-20 #367 @Bisaloo
Thank you for your work! I think the package has much improved.

The authors have addressed all my comment either by updating the package or making a receivable answer as to why they'd rather not, so I recommend approving this package.

I have submitted a PR with a couple of very minor changes. I hope you don't mind. It seems like it will save time to everyone to make the changes directly instead of copy/pasting the relevant lines here.

Two last minor comments:

- the first one is related to the `filterMatrix()` function. The `@return` value should be updated since it doesn't return a matrix but a data.frame. Maybe a name change is warranted as well(?)
- I think you did well by removing the prompt in `droptreedata.table()` since both Kari and I were confused by this and thought it was modifying in place. In the same vein, I recommend you update the message and leave out the 'ORIGINAL' adjective:

https://github.com/uyedaj/treedata.table/blob/3f308e6cda841ca2f4557b07c81f972ecf67b1bc/R/droptreedata.table.R#L73
