### 2020-08-24 #367 @cromanpa94
Thank you so much @Bisaloo! We just addressed all the remaining comments int the following commits:

https://github.com/uyedaj/treedata.table/commit/4917b4aa4d616bfa2b08189beae1266f0562f93e

We also included additional details on why to talk about Matrix in `filterMatrix()`. The documentation now distinguishes between character matrix (describing the features of organisms) and the class of the input (a data.frame).

We're happy to make any other changes if necessary!
