### 2020-08-25 1d27bac
rOpenSci review by Aug24th (#3)
* REVIEW:Type field deleted

* REVIEW: Title case

* REVIEW: LICENSE updated

* REVIEW:Multiline instructions removed

* Fix multiline as.treedata.table

* REVIEW:Booleans fixed in ifelse statements

* REVIEW:using inherits instead of class

* REVIEW:Fix seq_along

* REVIEW:message() instead of cat()

* REVIEW:paste0() within message() were removed

* REVIEW:name_column argument fixed for "detect"

* REVIEW:testthat updated to include correct column auto-detection

* REVIEW:Message printing the auto-detected column

* REVIEW:Changes based on style_pkg

* Remove 
[…]
