+++ R/extractVector.R
@@ -1,34 +1,37 @@
 #' Returning a named vector from a treedata.table object
 #'
 #' @param tdObject A treedata.table object
-#' @param ... The name of the column to select
+#' @param ... The name of the column or columns to select.
 #'
-#' @return A named vector
+#' @return A named vector or a list of named vectors
 #'
 #' @examples
 #'
 #' data(anolis)
-#' td <- as.treedata.table(tree=anolis$phy, data=anolis$dat)
-#' extractVector(td, "SVL") #extracts the named vector for SVL from the td object
-#'
+#' td <- as.treedata.table(tree = anolis$phy, data = anolis$dat)
+#' # extracts the named vector for SVL from the td object
+#' extractVector(td, "SVL")
+#' # extracts the named vector for SVL and ecomorph from the td object
+#' extractVector(td, "SVL", "ecomorph")
 #' @export
 
-extractVector <- function(tdObject, ...){
-  if(class(tdObject) != "treedata.table" ){
+extractVector <- function(tdObject, ...) {
+  if (!inherits(tdObject, c("treedata.table"))) {
     stop("Please use a class 'treedata.table' object \n")
   }
 
   dat <- tdObject$dat
-  args <- as.character(substitute(list(...)))[-1L]
-  arg_sub <- utils::type.convert(args)
-  if(is.numeric(arg_sub) | is.integer(arg_sub)) args <- arg_sub
-  vecs <- lapply(args,function(x) dat[[x]])
-  vecs <- if(class(tdObject$phy) =='phylo' ){
-    lapply(vecs, function(x) stats::setNames(x, tdObject$phy$tip.label)) }else{
-      lapply(vecs, function(x) stats::setNames(x, tdObject$phy[[1]]$tip.label))
-    }
-  if(length(vecs) == 1){
+  dots <- lazyeval::lazy_dots(...)
+  vecs <- lapply(list(...), function(x) dat[[x]])
+  vecs <- if (inherits(tdObject$phy, c("phylo"))) {
+    lapply(vecs, function(x) stats::setNames(x, tdObject$phy$tip.label))
+  } else {
+    lapply(vecs, function(x) stats::setNames(x, tdObject$phy[[1]]$tip.label))
+  }
+  if (length(vecs) == 1) {
     vecs <- vecs[[1]]
-  } else {names(vecs) <- args}
+  } else {
+    names(vecs) <- lapply(dots, function(x) x[[1]])
+  }
   return(vecs)
 }
