+++ tests/testthat/test.treedata.table.R
@@ -5,13 +5,15 @@ library(ape)
 data(anolis)
 td <- as.treedata.table(tree = anolis$phy, data = anolis$dat)
 tdt_output <-
-  tdt(td,
-      geiger::fitContinuous(
-        phy,
-        extractVector(td, SVL),
-        model = "BM",
-        ncores = 1
-      ))
+  tdt(
+    td,
+    geiger::fitContinuous(
+      phy,
+      extractVector(td, "SVL"),
+      model = "BM",
+      ncores = 1
+    )
+  )
 
 test_that("The resulting td object is of class treedata.table", {
   expect_is(td, "treedata.table")
@@ -22,14 +24,13 @@ test_that("Trees are the same", {
 })
 
 test_that("Species in tree and trait dataframe are the same", {
-
   expect_equal(td$phy$tip.label, td$dat$tip.label)
-
 })
 
 
 test_that("datasets are the same", {
-  expect_true(all(colnames(anolis$dat)[-1] == colnames(as.data.frame(td$dat))[-1]))
+  expect_true(all(colnames(anolis$dat)[-1] ==
+    colnames(as.data.frame(td$dat))[-1]))
 })
 
 
@@ -51,21 +52,386 @@ test_that(
   "The  number of rows is the same after filtering the original and tdt object
           under the same criteria",
   {
-    expect_equal(nrow(anolis$dat[anolis$dat$island == "Cuba" &
-                                   anolis$dat$ecomorph == "TG", ]),
-                 nrow(td[island == "Cuba" & ecomorph == "TG", ]$dat))
-
+    expect_equal(
+      nrow(anolis$dat[anolis$dat$island == "Cuba" &
+        anolis$dat$ecomorph == "TG", ]),
+      nrow(td[island == "Cuba" & ecomorph == "TG", ]$dat)
+    )
   }
 )
 
-test_that("[[ extracts a named character vector",{
+test_that("[[ extracts a named character vector", {
   expect_is(names(td[["SVL"]]), "character")
   expect_is(td[["SVL"]], "numeric")
 })
 
 
 
-test_that("phy and dat objects can be extracted correctly using pull.treedata.table",{
-  expect_is(pull.treedata.table(td, type = "phy"), "phylo")
-  expect_is(pull.treedata.table(td, type = "dat"), "data.table")
+test_that("phy and dat objects can be extracted correctly using
+          pulltreedata.table", {
+  expect_is(pulltreedata.table(td, type = "phy"), "phylo")
+  expect_is(pulltreedata.table(td, type = "dat"), "data.table")
+})
+
+test_that("Column containing tip labs can be correctly detected", {
+  tre <- anolis$phy
+  dat1 <- anolis$dat
+  dat2 <- dat1[, sample(ncol(dat1), ncol(dat1))]
+  td1 <- as.treedata.table(tre, dat1)
+  td2 <- as.treedata.table(tre, dat2)
+  expect_equal(td1$phy, td2$phy)
+})
+
+test_that("Find the correct number of discrete/continuous characters in the
+          anolis dataset", {
+  expect_equal(
+    detectCharacterType(anolis$dat[, 1]),
+    detectAllCharacters(anolis$dat)[1]
+  )
+})
+
+
+
+test_that("head() returns a data.table object", {
+  expect_is(head(td), "data.table")
+})
+
+test_that("tail() returns a data.table object", {
+  expect_is(tail(td), "data.table")
+})
+
+test_that("Error is shown when tips with different tip labels are used", {
+  anolis2 <- anolis$phy
+  anolis2$tip.label[1] <- "NAA"
+  tree2 <- list(anolis$phy, anolis2)
+  class(tree2) <- "multiPhylo"
[…]
