### 2020-08-25 #367 @jooolia
Great! Thanks @cromanpa94 , @wrightaprilm  and @uyedaj for the submission and @karinorman and @Bisaloo for the reviews. I am a guest editor so I need to get help creating the github team that will allow us to transfer the repository over to ropensci and the next steps. I will follow up soon. Thanks, Julia
