### 2020-08-26 d93acd2
a couple wording tweaks
