### 2020-08-31 #367 @jooolia
Thanks again @uyedaj, @wrightaprilm,and @cromanpa94  for submitting and @karinorman and @Bisaloo for your reviews! :)

To-dos:
- [ ] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  Maelle Salmon has invited you to a team that should allow you to do so.  Let us know when you've done this and I can let her know to make you admins once you do.
    * replace your package docs URL with `https://docs.ropensci.org/package_name`

Thanks a lot for your contribution and your patience with the process during this difficult time.

Cheers, Julia
