+++ R/anolis.R
@@ -1,6 +1,8 @@
 #' @name anolis
 #' @title Anole data
-#' @description Anole data for treedata.table functions
+#' @description Anole data for treedata.table functions. Many
+#' of the traits (e.g. awesomeness, hostility) in this dataset
+#' retrieved from treeplyr are based on random numbers.
 #' @author Luke Harmon
 #' @docType data
 #' @usage data(anolis)
