### 2020-09-07 dd3749a
Vignettes updated
