+++ vignettes/treedata.table_spanish.Rmd
@@ -1,0 +1,154 @@
+---
+title: "Breve Introducción al Paquete treedata.table"
+author: "Cristian Roman-Palacios, April Wright, Josef Uyeda"
+date: "09/06/2020"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Getting Started With The treedata.table Package}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+## Breve Introducción al Paquete `treedata.table`
+
+El paquete `treedata.table` tiene como objetivo permitir a investigadores acceder y manipular datos filogenéticos usando herramientas del paquete `data.table`. `data.table` tiene diferentes funciones para manipular rápidamente datos en una forma eficiente.
+
+El primer paso para usar `treedata.table` consiste en crear un objeto `treedata.table`. El objeto `treedata.table` empareja los tip.labels de la filogenia con una columna en el `data.frame` que contiene los nombres de los taxones. Este paso inicial permite la manipulación subsecuente y coordinada de los datos en el árbol y la matriz de caracteres dentro de `treedata.table`.
+
+Para este tutorial vamos a usar los datos de Anolis creados en `treeplyr`. Estos caracteres fueron generados aleatoriamente. Es importante resaltar tres aspectos. Primero, el árbol tiene que ser en formato `phylo` (o `multiPhylo` en caso de múltiples arboles). Segundo, la matriz de caracteres tiene que están en formato `data.frame`. Tercero, la matriz de caracteres tiene que contener una columna con los nombres de los taxones coincidiendo con los tip.labels del árbol (o arboles). 
+
+El objeto `treedata.table` se crea usando la función `as.treedata.table`.
+
+```{r}
+library(ape)
+library(treedata.table)
+
+# Cargamos los datos del ejemplo
+data(anolis)
+# Creamos el objecto treedata.table con as.treedata.table
+td <- as.treedata.table(tree = anolis$phy, data = anolis$dat)
+```
+
+Podemos revisar el objeto resultante simplemente llamando el nombre del objeto en la consola. La matriz de caracteres, antes en `data.frame`, es ahora un `data.table`
+
+```{r}
+td
+```
+
+Adicionalmente, la matriz de caracteres en el nuevo formato `data.table` ha sido re-ordenado para tener las filas en el mismo orden que los tip.labels en el árbol.
+
+
+```{r}
+td$phy$tip.label == td$dat$tip.label
+```
+
+
+## Manipulando Datos
+
+La matriz de datos en el objeto treedata.table puede ser indexada como cualquier otro objeto `data.table`. Por ejemplo, podemos hacer lo siguiente para extraer la columna con información de longitud hocico-cloaca (SVL) para cada especie.
+
+```{r pressure, echo=TRUE}
+td$dat[,'SVL']
+
+```
+
+También podemos usar los brackets dobles para extraer directamente la misma columna como un vector con nombres.
+
+```{r}
+td[["SVL"]]
+```
+
+El mismo resultado puede también ser logrado usando la función `extractVector`. El producto de `extractVector` como los brackets dobles es un vector con nombres.
+
+
+```{r}
+extractVector(td, 'SVL')
+```
+
+Múltiples columnas pueden tambien ser extraídas usando `extractVector`.
+
+```{r}
+extractVector(td, 'SVL','ecomorph')
+```
+
[…]
