+++ man/anolis.Rd
@@ -9,7 +9,9 @@
 data(anolis)
 }
 \description{
-Anole data for treedata.table functions
+Anole data for treedata.table functions. Many
+of the traits (e.g. awesomeness, hostility) in this dataset
+retrieved from treeplyr are based on random numbers.
 }
 \author{
 Luke Harmon
