### 2020-09-07 #367 @maelle
@cromanpa94 @uyedaj I re-sent an invitation to the ropensci GitHub organization. @wrightaprilm  was already in the team and org.

@stefaniebutland I think you mean the slack workspace, which is useful in any case!
