+++ vignettes/1.1_treedata.table_intro_english.Rmd
@@ -146,177 +146,11 @@ The table
 df
 ```
 
-and the respective tree
+and the corresponding tree
 
 ```{r}
 tree
 ```
-
-
-# Additional functions for manipulating data
-
-`treedata.table` includes additional functions that allow the identification of `discrete` and `continuous` character in a given dataset. For instance, the `detectCharacterType()` function can be used to examine whether `SVL` is `discrete` or `continuous`: 
-
-```{r}
-detectCharacterType(anolis$dat$SVL)
-```
-
-We can further examine the type of characters we have in our dataset by using the `detectAllCharacters()` function:
-
-```{r}
-detectAllCharacters(anolis$dat)
-```
-
-Summarizing this information in a table, we get:
-
-```{r}
-cbind.data.frame(character=colnames(anolis$dat),type=detectAllCharacters(anolis$dat))
-```
-
-Finally, we can use the `filterMatrix()` function to subset our dataset to only a certain type of characters. For instance, let's extract all discrete characters in the *Anolis* dataset:
-
-```{r}
-filterMatrix(anolis$dat, "discrete")
-```
-
-Two additional functions in `treedata.table` are designed to examine and modify column and row names in any dataset. For instance, we can ask if the *Anolis* dataset has column names:
-
-```{r}
-hasNames(anolis$dat, "col")
-```
-
-It does have column names. Let's just remove the column names and check if `hasNames()` can detect this change. Here's our new dataset:
-
-```{r}
-data=anolis$dat
-colnames(data)<-NULL
-head(data,2)
-```
-
-Let's run `hasNames()` on our new dataset:
-
-```{r}
-hasNames(data, "col")
-```
-
-Now, we can create new column names using the `forceNames()` function:
-
-```{r}
-data <- forceNames(data, "col")
-```
-
-The new dataset, with column names (n1...), looks like this:
-
-```{r}
-head(data,2)
-```
-
-And we can finally ask whether our new dataset actually have column names by running the `hasNames()` function again:
-
-
-```{r}
-hasNames(data, "col")
-```
-
-
-We can apply the same procedure for columns (`col`), rows (`row`) or both (`rowcol`).
-
-
-# Advanced methods
-
-## Working with multiphylo objects
-
-`treedata.table` further allows the matching of multiple phylogenies (`multiPhylo`) against a single dataset (`data.frame`). Below, we modified the anole dataset to explain the extended functionality of `treedata.table` with `multiPhylo` objects.
-
-
-We first create a `multiPhylo` object including only two `phylo` objects. Users can provide any number of `phylo` objects within the `multiPhylo` object. However, trees should only differ in their topology. In other words, all trees must have the same tip labels. Nevertheless, both the provided `multiPhylo` and `data.frame` should partially overlap.
-
-```{r}
-trees<-list(anolis$phy,anolis$phy)
-class(trees) <- "multiPhylo"
-trees
-```
-
-Now, we create our treedata.table object by combining the trait data (`data.frame`) and the newly generated `multiPhylo` object.
-
-```{r}
-td <- as.treedata.table(tree=trees, data=anolis$dat)
-```
-
-The resulting `td` object now returns a `multiPhylo` object under `phy` 
-
-```{r}
-class(td$phy);td$phy
-```
-
-Please note that all the basic `treedata.table` functions highlighted above for `phylo` objects are still functional when `treedata.table` objects include `multiPhylo` objects.
-
-```{r}
-td[, head(.SD, 1), by = "ecomorph"]
-```
-
[…]
