+++ pkgdown.yml
@@ -5,6 +5,7 @@ articles:
   1.1_treedata.table_intro_english: 1.1_treedata.table_intro_english.html
   1.2_treedata.table_intro_spanish: 1.2_treedata.table_intro_spanish.html
   2_multiphylo_treedata.table: 2_multiphylo_treedata.table.html
-  3_AdditionalFunctions_treedata.table: 3_AdditionalFunctions_treedata.table.html
-last_built: 2020-09-07T06:50Z
+  3_PartialMatching: 3_PartialMatching.html
+  4_AdditionalFunctions_treedata.table: 4_AdditionalFunctions_treedata.table.html
+last_built: 2020-09-07T06:59Z
 
