### 2020-09-07 7d42f4e
Update README.md
Updated install address to ropensci
