+++ pkgdown.yml
@@ -7,5 +7,5 @@ articles:
   B_multiphylo_treedata.table: B_multiphylo_treedata.table.html
   C_PartialMatching: C_PartialMatching.html
   D_AdditionalFunctions_treedata.table: D_AdditionalFunctions_treedata.table.html
-last_built: 2020-09-07T07:26Z
+last_built: 2020-09-07T15:09Z
 
