+++ DESCRIPTION
@@ -26,8 +26,8 @@ Authors@R: c(
           family = "Norman",
           role = c("rev"))
   )
-URL: https://github.com/uyedaj/treedata.table/
-BugReports: https://github.com/uyedaj/treedata.table/issues
+URL: https://github.com/ropensci/treedata.table/
+BugReports: https://github.com/ropensci/treedata.table/issues
 Description: This package combines trait data and a phylogenetic tree (or trees) into a 
     single object of class treedata.table. The resulting object can be easily 
     manipulated to simultaneously change the trait- and tree-level sampling.
