### 2020-09-07 #367 @uyedaj
Thank you @maelle and @jooolia, I've transferred the repo to ropensci's github organization and need to be made an admin now to do the next steps.
