+++ pkgdown.yml
@@ -1,11 +1,6 @@
 pandoc: 2.7.3
-pkgdown: 1.5.1
+pkgdown: 1.6.0
 pkgdown_sha: ~
-articles:
-  AA_treedata.table_intro_english: AA_treedata.table_intro_english.html
-  AB_treedata.table_intro_spanish: AB_treedata.table_intro_spanish.html
-  B_multiphylo_treedata.table: B_multiphylo_treedata.table.html
-  C_PartialMatching: C_PartialMatching.html
-  D_AdditionalFunctions_treedata.table: D_AdditionalFunctions_treedata.table.html
-last_built: 2020-09-07T15:14Z
+articles: []
+last_built: 2020-09-08T22:11Z
 
