### 2020-09-08 1af6be5
updated description url to new ropensci pkgdown site
