### 2020-09-08 39cc06b
Minor changes to the Additional resources
