### 2020-09-11 #367 @uyedaj
I hope we have completed all of these satisfactorily:

> To-dos:
>
> * [x]  Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  Maelle Salmon has invited you to a team that should allow you to do so.  Let us know when you've done this and I can let her know to make you admins once you do.
> * [x]  Fix all links to the GitHub repo to point to the repo under the ropensci organization.
> * [x]  If you already had a `pkgdown` website **and are ok relying only on [rOpenSci central docs building and branding](https://devguide.ropensci.org/ci.html#even-more-ci-ropensci-docs)**,
>
>   * deactivate the automatic deployment you might have set up
>   * remove styling tweaks from your pkgdown config but keep that config file
>   * replace the whole current `pkgdown` website with a [redirecting page](https://devguide.ropensci.org/#redirect)
>   * replace your package docs URL with `https://docs.ropensci.org/package_name`
>   * In addition, in your DESCRIPTION file, include the docs link in the `URL` field alongside the link to the GitHub repository, e.g.: `URL: https://docs.ropensci.org/foobar (website) https://github.com/ropensci/foobar`
> * [x]  Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`.
> * [x]  We're starting to roll out software metadata files to all ropensci packages via the Codemeta initiative, see https://github.com/ropensci/codemetar/#codemetar for how to include it in your package, after installing the package - should be easy as running codemetar::write_codemeta() in the root of your package.
>
>  [x] Should you want to acknowledge your reviewers in your package DESCRIPTION, you can do so by making them `"rev"`-type contributors in the `Authors@R` field (with their consent). More info on this [here](https://devguide.ropensci.org/building.html#authorship).

Thank you for your help!
