### 2020-09-17 #367 @jooolia
Great thanks @uyedaj ! It looks good and I will close this issue. I think that you can also close your issue related to this review in your repo.
If you or @wrightaprilm or @cromanpa94  are interested in writing a blog post to promote your package that is a possibility, but it is optional and up to you.
Thanks again for the package and the nice submission and review process.
Cheers, Julia
